==> code/01-Prep_BICS_Data.Rmd <==
---
title: "Comparison"
author: "Christopher Soria"
date: "2024-01-24"
output: html_document
---
```{r}
library(dplyr)
library(tidyverse)
library(data.table)
library(readxl)
library(stringr)
```

```{r}
ayesha_path <- 'insert_path_here'
dennis_path <- 'insert_path_here'
chris_path <- '/Users/chrissoria/Documents/Research/BICS_Political_Polarization'
```

switch between manual and combined
```{r}
file_type <- "manual"
setwd(chris_path) #user enters file path here

path_set <- getwd()
print(path_set)
```

some ZIP Codes, such as "unique" ZIPs that are often tied to an internal mail-routing mechanism (such as UC Berkeley which has its own internal ZIP code and mail-routing system) will have very low ratios).
As of 12/23 I'm realzing that I'll need to start with a complete list of all ZIP codes (from USPS) before I start this so I can get a good sense of completeness.

Below, I'm reading in that full list of ZIP codes
```{r}
ZIP_to_CD_df <- read_excel(paste0(path_set,'/Data/ZIP_CD_062020.xlsx')) %>%
  rename(
    CD_RES_RATIO = RES_RATIO,
    CD_BUS_RATIO = BUS_RATIO,
    CD_OTH_RATIO = OTH_RATIO,
    CD_TOT_RATIO = TOT_RATIO
  ) %>%
  mutate(ZIP = str_pad(as.character(ZIP), width = 5, pad = "0"))

print(ZIP_to_CD_df)
```
```{r}
# Read the Excel file and ensure all columns are read as text otherwise leading zeroes in ZIP will be cut
FULL_ZIPS <- read_excel(paste0(path_set, '/Data/Full_Zips.xlsx'), col_types = "text")

FULL_ZIPS <- FULL_ZIPS %>%
  rename(
    ZIP = `DELIVERY ZIPCODE`,
    USPS_STATE = `PHYSICAL STATE`
  )

FULL_ZIPS <- FULL_ZIPS[, c("ZIP", "USPS_STATE", "TYPE")]

# Remove duplicate ZIP codes and clean the data
FULL_ZIPS <- FULL_ZIPS %>%
  distinct(ZIP, .keep_all = TRUE) %>%
  mutate(ZIP = str_pad(ZIP, width = 5, pad = "0")) %>%
  filter(!is.na(USPS_STATE)) %>%
  filter(!USPS_STATE %in% c("VI", "PR", "MH", "MP", "FM", "GU")) %>%
  filter(!TYPE %in% c("Other", "Unique")) %>%
  arrange(ZIP) %>%
  ungroup()

print(nrow(FULL_ZIPS))
print(head(FULL_ZIPS))
```
After merging, I find that many of the Alaska ZIP codes aren't getting matched. Since there's only one CD in Alaska, I'll fill in any Alaska state ZIP with 0200.
This data will not match PO boxes. Most non matches are coming from PO boxes.
```{r}
ZIP_FEATURES <- left_join(FULL_ZIPS, ZIP_to_CD_df, by = "ZIP")

ZIP_FEATURES <- ZIP_FEATURES %>%
  mutate(
    CD = ifelse(USPS_STATE == "AK", "0200", CD),
    CD_TOT_RATIO = ifelse(USPS_STATE == "AK", 1.0, CD_TOT_RATIO),
    CD = ifelse(ZIP == "15253", "4212", CD)
  )

# the resulting dataframe is a ZIP code matches to CD
print(ZIP_FEATURES)
```
For now, I will assign a ZIP code to whatever whatever County it falls mostly under residential
We will run this as an outer match because then we can keep zip codes from the right side that aren't in the left side
```{r}
ZIP_to_County_df <- read_excel(paste0(path_set, '/Data/ZIP_COUNTY_062020_HUD.xlsx'), col_types = "text")

# Ensure COUNTY and ZIP columns are strings with zero padding
ZIP_to_County_df <- ZIP_to_County_df %>%
  mutate(
    COUNTY = str_pad(COUNTY, width = 5, pad = "0"),
    ZIP = str_pad(ZIP, width = 5, pad = "0")
  )

print(head(ZIP_to_County_df))

ZIP_FEATURES <- full_join(ZIP_FEATURES, ZIP_to_County_df, by = "ZIP")

ZIP_FEATURES <- ZIP_FEATURES %>%
  arrange(desc(RES_RATIO)) %>%
  distinct(ZIP, .keep_all = TRUE)

ZIP_to_County_df <- ZIP_to_County_df %>%
  rename(
    COUNTY_RES_RATIO = RES_RATIO,
    COUNTY_BUS_RATIO = BUS_RATIO,
    COUNTY_OTH_RATIO = OTH_RATIO,
    COUNTY_TOT_RATIO = TOT_RATIO
  )

#the resulting DF now matches zip codes to county
print(head(ZIP_FEATURES))

# Count the number of missing values in the COUNTY column and calculate the percentage
missing_count <- sum(is.na(ZIP_FEATURES$COUNTY))
missing_percentage <- mean(is.na(ZIP_FEATURES$COUNTY)) * 100

print(missing_count)  
print(missing_percentage) # High level of match
print(nrow(ZIP_FEATURES))
```
```{r}
CD_df <- read_csv(paste0(path_set, '/data/House_Reps_District.csv'), col_types = cols(
  FIPS_Code = col_character(),
  District = col_character(),
  CD = col_character(),
  .default = col_guess()  # Guess the types of the remaining columns
))
CD_df
```


```{r}
# Ensure the 'CD' column is a character in both dataframes
ZIP_FEATURES <- ZIP_FEATURES %>%
  mutate(CD = as.character(CD))

CD_df <- CD_df %>%
  mutate(CD = as.character(CD))

# Merge ZIP_FEATURES with CD_df using a left join on 'CD'
ZIP_FEATURES <- left_join(ZIP_FEATURES, CD_df, by = "CD")

# Rename columns in ZIP_FEATURES
ZIP_FEATURES <- ZIP_FEATURES %>%
  rename(
    COUNTY_FIPS = COUNTY,
    CONGRESSPERSON = Name,
    STATE = State,
    CONGRESSPERSON_PARTY = Party
  )

cols_to_move <- colnames(ZIP_FEATURES)[3:5]
other_cols <- colnames(ZIP_FEATURES)[!(colnames(ZIP_FEATURES) %in% cols_to_move)]
ZIP_FEATURES <- ZIP_FEATURES[, c(other_cols, cols_to_move)]

ZIP_FEATURES
```
```{r}
mask_mandates_count <- read_excel(paste0(path_set, '/data/county_mask_mandate_data.xlsx'), col_types = c("text"))

mask_mandates_count <- mask_mandates_count %>%
  select(county_fips, county_conditions, state_conditions)

mask_mandates_count <- mask_mandates_count %>%
  mutate(county_fips = str_pad(county_fips, width = 5, pad = "0"))

mask_mandates_count <- mask_mandates_count %>%
  mutate(county_conditions = ifelse(is.na(county_conditions) & !is.na(state_conditions), state_conditions, county_conditions)) %>%
  mutate(county_conditions = ifelse(is.na(county_conditions), "no mandate", county_conditions))

mask_mandates_count <- mask_mandates_count %>%
  rename(COUNTY_FIPS = county_fips)

print(head(mask_mandates_count))

ZIP_FEATURES <- left_join(ZIP_FEATURES, mask_mandates_count, by = "COUNTY_FIPS")

ZIP_FEATURES
```
I have a pretty decent match rate of 95%. I believe most of these will be from DC.

```{r}
num_missing_congressperson <- sum(is.na(ZIP_FEATURES$CONGRESSPERSON))
perc_missing_congressperson <- mean(is.na(ZIP_FEATURES$CONGRESSPERSON)) * 100

cat("Number of missing values in CONGRESSPERSON:", num_missing_congressperson, "\n")
cat("Percentage of missing values in CONGRESSPERSON:", perc_missing_congressperson, "%\n")

num_missing_county_conditions <- sum(is.na(ZIP_FEATURES$county_conditions))
perc_missing_county_conditions <- mean(is.na(ZIP_FEATURES$county_conditions)) * 100

cat("Number of missing values in county_conditions:", num_missing_county_conditions, "\n")
cat("Percentage of missing values in county_conditions:", perc_missing_county_conditions, "%\n")
```
Below, we need to add a leading "0" to the FIPS code (FIPS codes are all 5 digits)
These are cumulative mortality numbers (not weekly)
```{r}
usafacts_deaths <- read_csv(paste0(path_set, '/data/Partisanship and Health Behavior/Data/covid_deaths_usafacts.csv'), col_types = cols(
  countyFIPS = col_character()
))

usafacts_deaths <- usafacts_deaths %>%
  mutate(countyFIPS = str_pad(countyFIPS, width = 5, pad = "0"))
```
For this preliminary analysis, I will only use wave 2,4,6.
Each wave's data was collected on 2020-06-17 - 2020-06-23, 2020-11-29 - 2020-12-16, 2021-05-12 - 2021-05-25.
I will take covid mortality from the midpoint of the wave range.
Wave 2 = 6/18-24. 2020
Wave 3 = 9/12-20, 2020
Wave 4 = 12/1-7, 2020
Wave 5 = 2/9-14, 2021
Wave 6 = 5/14-25, 2021

```{r}
usafacts_deaths <- read_csv(paste0(path_set, '/Data/Partisanship and Health Behavior/Data/covid_deaths_usafacts.csv'), col_types = cols(
  countyFIPS = col_character()
))

usafacts_deaths <- usafacts_deaths %>%
  mutate(countyFIPS = str_pad(countyFIPS, width = 5, pad = "0"))

# Define the columns to keep
columns_to_keep <- c('countyFIPS', 'County Name',
                     '2020-06-01', '2020-06-15', '2020-06-13', '2020-06-20', # w2
                     '2020-09-01', '2020-09-16', '2020-09-09', '2020-09-05', # w3
                     '2020-12-05', '2020-12-04', '2020-12-01', '2020-11-28', # w4
                     '2021-02-12', '2021-02-10', '2021-02-05', '2021-01-29', # w5
                     '2021-05-19', '2021-05-17', '2021-05-12', '2021-05-05') # w6

usafacts_deaths <- usafacts_deaths %>%
  select(all_of(columns_to_keep))

new_column_names <- ifelse(seq_along(columns_to_keep) <= 2, columns_to_keep, 
                           paste0('COUNTY_COV_DEATHS_', gsub('-', '_', columns_to_keep)))

colnames(usafacts_deaths) <- new_column_names

usafacts_deaths <- usafacts_deaths %>%
  rename(COUNTY_FIPS = countyFIPS)

print(head(usafacts_deaths))
```
Nearly all ZIP codes get a match (just over 1 percent don't get matched).
If we end up going back to mortality rates (as of now (01/14/24) we're using prevalence rates.
```{r}
ZIP_FEATURES <- left_join(ZIP_FEATURES, usafacts_deaths, by = "COUNTY_FIPS")

num_missing_cov_deaths <- sum(is.na(ZIP_FEATURES$COUNTY_COV_DEATHS_2020_06_01))
perc_missing_cov_deaths <- mean(is.na(ZIP_FEATURES$COUNTY_COV_DEATHS_2020_06_01)) * 100

cat("Number of missing values in COUNTY_COV_DEATHS_2020_06_01:", num_missing_cov_deaths, "\n")
cat("Percentage of missing values in COUNTY_COV_DEATHS_2020_06_01:", perc_missing_cov_deaths, "%\n")
```
```{r}
census_pop <- read_csv(paste0(path_set, '/Data/Partisanship and Health Behavior/Data/co-est2021-alldata.csv'), 
                       col_types = cols(
                         STATE = col_character(),
                         COUNTY = col_character(),
                         .default = col_guess()
                       ), 
                       locale = locale(encoding = 'ISO-8859-1'))

census_pop_FIPS <- read_excel(paste0(path_set, '/Data/Partisanship and Health Behavior/Data/all-geocodes-v2020.xlsx'), col_types = c("text"))

census_pop_FIPS <- census_pop_FIPS %>%
  rename(CTYNAME = `Area Name (including legal/statistical area description)`)

census_pop <- census_pop %>%
  mutate(COUNTY_FIPS = str_c(STATE, COUNTY))

State_FIPS_Crosswalk <- census_pop %>%
  select(STATE, STNAME) %>%
  mutate(STATE_FIPS = as.character(STATE)) %>%
  mutate(STATE = STNAME) %>%
  select(-STNAME) %>%
  distinct(STATE, .keep_all = TRUE) %>%
  filter(!is.na(STATE))

write_csv(State_FIPS_Crosswalk, paste0(path_set, '/data/STATE_FIPS_CROSSWALK.csv'))
write_csv(State_FIPS_Crosswalk, paste0(path_set, '/Data/Partisanship and Health Behavior/Data/STATE_FIPS_CROSSWALK.csv'))

columns_to_keep <- c('STATE', 'POPESTIMATE2020', 'DEATHS2020', 'POPESTIMATE2021', 'DEATHS2021', 'COUNTY_FIPS')
census_pop <- census_pop %>%
  select(all_of(columns_to_keep)) %>%
  rename(STATE_FIPS = STATE)

census_pop
State_FIPS_Crosswalk
```
A crude mortality rate, not standardized or age adjusted, for this preliminary analysis
I manually insert Puerto Rico' State FIPS code and District of Columbia to make this a clean merge across the board.
```{r}
ZIP_FEATURES <- left_join(ZIP_FEATURES, census_pop, by = "COUNTY_FIPS")

ZIP_FEATURES <- ZIP_FEATURES %>%
  mutate(
    COV_County_CMR_2020_06_20 = (COUNTY_COV_DEATHS_2020_06_20 / POPESTIMATE2020) * 1000,
    COV_County_CMR_2020_12_09 = (COUNTY_COV_DEATHS_2020_12_05 / POPESTIMATE2020) * 1000,
    COV_County_CMR_2021_05_19 = (COUNTY_COV_DEATHS_2021_05_19 / POPESTIMATE2021) * 1000,
    STATE_FIPS = if_else(STATE == "Puerto Rico", "72", STATE_FIPS)
  )

num_missing_cmr_2020_06_20 <- sum(is.na(ZIP_FEATURES$COV_County_CMR_2020_06_20))
perc_missing_cmr_2020_06_20 <- mean(is.na(ZIP_FEATURES$COV_County_CMR_2020_06_20)) * 100

cat("Number of missing values in COV_County_CMR_2020_06_20:", num_missing_cmr_2020_06_20, "\n")
cat("Percentage of missing values in COV_County_CMR_2020_06_20:", perc_missing_cmr_2020_06_20, "%\n")
```
We have decided that prevalence is a better measure for predicting contact behavior than mortality
I want these lags
Wave 2 (6/18-24, 2020), midpoint: 06/21/2020, lag1: 06/14/2020, lag3: 06/07/2020
Wave 3 (9/12-20, 2020), midpoint: 09/16/2020, lag1: 09/09/2020, lag3: 09/02/2020
Wave 4 (12/1-7, 2020), midpoint: 12/04/2020, lag1: 11/27/2020, lag3: 11/20/2020
Wave 5 (2/9-14, 2021), midpoint: 02/11/2021, lag: 02/04/2021, lag3: 01/28/2021
Wave 6 (5/14-25, 2021), midpoint: 05/19/2021, lag: 05/12/2021, lag3: 05/05/2021
```{r}
# Read and process prevalence_mid_2020 data
prevalence_mid_2020 <- read_csv(paste0(path_set, '/Data/Partisanship and Health Behavior/Data/JH_Prevalence/time_series_covid19_confirmed_US_mid_2020.csv'), 
                                col_types = cols(.default = col_guess()))

columns_to_keep <- c('FIPS', '6/21/20', '6/18/20', '6/14/20', '6/7/20')

prevalence_mid_2020 <- prevalence_mid_2020 %>%
  filter(iso2 == 'US') %>%
  mutate(FIPS = as.character(FIPS),
         FIPS = if_else(!is.na(FIPS) & str_detect(FIPS, "\\."), str_remove(FIPS, "0+$|\\.$"), FIPS),
         FIPS = str_pad(FIPS, width = 5, pad = "0")) %>%
  select(all_of(columns_to_keep))

# Calculate wave_2_incidence and wave_2_inc_week_lag
prevalence_mid_2020 <- prevalence_mid_2020 %>%
  mutate(wave_2_incidence = as.numeric(`6/21/20`) - as.numeric(`6/14/20`),
         wave_2_inc_week_lag = as.numeric(`6/14/20`) - as.numeric(`6/7/20`))

write_csv(prevalence_mid_2020, paste0(path_set, '/data/wave_2_incidence.csv'))

# Read and process prevalence_late_2020 data
prevalence_late_2020 <- read_csv(paste0(path_set, '/Data/Partisanship and Health Behavior/Data/JH_Prevalence/time_series_covid19_confirmed_US_late_2020.csv'), 
                                col_types = cols(.default = col_guess()))

columns_to_keep <- c('FIPS', '9/16/20', '9/12/20', '9/9/20', '9/2/20')

prevalence_late_2020 <- prevalence_late_2020 %>%
  filter(iso2 == 'US') %>%
  mutate(FIPS = as.character(FIPS),
         FIPS = if_else(!is.na(FIPS) & str_detect(FIPS, "\\."), str_remove(FIPS, "0+$|\\.$"), FIPS),
         FIPS = str_pad(FIPS, width = 5, pad = "0")) %>%
  select(all_of(columns_to_keep))

# Calculate wave_3_incidence and wave_3_inc_week_lag
prevalence_late_2020 <- prevalence_late_2020 %>%
  mutate(wave_3_incidence = `9/16/20` - `9/9/20`,
         wave_3_inc_week_lag = `9/9/20` - `9/2/20`)

write_csv(prevalence_late_2020, paste0(path_set, '/data/wave_3_incidence.csv'))

# Read and process prevalence_early_2021 data
prevalence_early_2021 <- read_csv(paste0(path_set, '/Data/Partisanship and Health Behavior/Data/JH_Prevalence/time_series_covid19_confirmed_US_early_2021.csv'), 
                                col_types = cols(.default = col_guess()))

columns_to_keep <- c('FIPS', '12/4/20', '12/1/20', '11/27/20', '11/20/20')

prevalence_early_2021 <- prevalence_early_2021 %>%
  filter(iso2 == 'US') %>%
  mutate(FIPS = as.character(FIPS),
         FIPS = if_else(!is.na(FIPS) & str_detect(FIPS, "\\."), str_remove(FIPS, "0+$|\\.$"), FIPS),
         FIPS = str_pad(FIPS, width = 5, pad = "0")) %>%
  select(all_of(columns_to_keep))

# Calculate wave_4_incidence and wave_4_inc_week_lag
prevalence_early_2021 <- prevalence_early_2021 %>%
  mutate(wave_4_incidence = `12/4/20` - `11/27/20`,
         wave_4_inc_week_lag = `11/27/20` - `11/20/20`)

write_csv(prevalence_early_2021, paste0(path_set, '/data/wave_4_incidence.csv'))

# Read and process prevalence_mid_2021 data
prevalence_mid_2021 <- read_csv(paste0(path_set, '/Data/Partisanship and Health Behavior/Data/JH_Prevalence/time_series_covid19_confirmed_US_mid_2021.csv'), 
                                col_types = cols(.default = col_guess()))

columns_to_keep <- c('FIPS', '2/11/21', '2/9/21', '2/4/21', '1/28/21')

prevalence_mid_2021 <- prevalence_mid_2021 %>%
  filter(iso2 == 'US') %>%
  mutate(FIPS = as.character(FIPS),
         FIPS = if_else(!is.na(FIPS) & str_detect(FIPS, "\\."), str_remove(FIPS, "0+$|\\.$"), FIPS),
         FIPS = str_pad(FIPS, width = 5, pad = "0")) %>%
  select(all_of(columns_to_keep))

# Calculate wave_5_incidence and wave_5_inc_week_lag
prevalence_mid_2021 <- prevalence_mid_2021 %>%
  mutate(wave_5_incidence = `2/11/21` - `2/4/21`,
         wave_5_inc_week_lag = `2/4/21` - `1/28/21`)

write_csv(prevalence_mid_2021, paste0(path_set, '/data/wave_5_incidence.csv'))

# Read and process prevalence_mid_2021_2 data
prevalence_mid_2021_2 <- read_csv(paste0(path_set, '/Data/Partisanship and Health Behavior/Data/JH_Prevalence/time_series_covid19_confirmed_US_mid_2021.csv'), 
                                col_types = cols(.default = col_guess()))

columns_to_keep <- c('FIPS', '5/19/21', '5/14/21', '5/12/21', '5/5/21')

prevalence_mid_2021_2 <- prevalence_mid_2021_2 %>%
  filter(iso2 == 'US') %>%
  mutate(FIPS = as.character(FIPS),
         FIPS = if_else(!is.na(FIPS) & str_detect(FIPS, "\\."), str_remove(FIPS, "0+$|\\.$"), FIPS),
         FIPS = str_pad(FIPS, width = 5, pad = "0")) %>%
  select(all_of(columns_to_keep))

# Calculate wave_6_incidence and wave_6_inc_week_lag
prevalence_mid_2021_2 <- prevalence_mid_2021_2 %>%
  mutate(wave_6_incidence = `5/19/21` - `5/12/21`,
         wave_6_inc_week_lag = `5/12/21` - `5/5/21`)

write_csv(prevalence_mid_2021_2, paste0(path_set, '/data/wave_6_incidence.csv'))

# Combine all prevalence data into one dataframe
prevalence <- prevalence_mid_2020 %>%
  left_join(prevalence_late_2020, by = "FIPS") %>%
  left_join(prevalence_early_2021, by = "FIPS") %>%
  left_join(prevalence_mid_2021, by = "FIPS") %>%
  left_join(prevalence_mid_2021_2, by = "FIPS")

# Rename columns to prepend 'prev_' to all columns except the first (FIPS)
colnames(prevalence) <- ifelse(seq_along(colnames(prevalence)) == 1, colnames(prevalence), paste0('prev_', colnames(prevalence)))

# Rename 'FIPS' to 'COUNTY_FIPS'
prevalence <- prevalence %>%
  rename(COUNTY_FIPS = FIPS) %>% 
  filter(!is.na(COUNTY_FIPS))

prevalence
```
Next, I want to create the weekly incidence for each key date

#06/21/2020 - 06/14/2020 prevalence['wave_2_incidence'] = prevalence['prev_6/21/20'] - prevalence['prev_6/14/20'] #09/16/2020 - 09/09/2020 prevalence['wave_3_incidence'] = prevalence['prev_9/16/20'] - prevalence['prev_9/9/20'] #12/04/2020 - 11/27/2020 prevalence['wave_4_incidence'] = prevalence['prev_12/4/20'] - prevalence['prev_11/27/20'] #02/11/2021 - 02/04/2021 prevalence['wave_5_incidence'] = prevalence['prev_2/11/21'] - prevalence['prev_2/4/21'] #05/19/2021 - 05/12/2021 prevalence['wave_6_incidence'] = prevalence['prev_5/19/21'] - prevalence['prev_5/12/21']

below we turn these figures into rates based on population
```{r}
ZIP_FEATURES <- left_join(ZIP_FEATURES, prevalence, by = "COUNTY_FIPS")

# Loop through the columns of ZIP_FEATURES
for (col in colnames(ZIP_FEATURES)) {
  # Check if the column name starts with 'prev_'
  if (startsWith(col, 'prev_')) {
    # Extract the date part of the column name
    date_part <- str_replace(col, 'prev_', '')
    # Perform the calculation and create a new column
    ZIP_FEATURES <- ZIP_FEATURES %>%
      mutate(!!paste0('rate_prev_', date_part) := (get(col) / POPESTIMATE2020) * 1000)
  }
}

# Drop the prevalence columns from ZIP_FEATURES
prev_columns <- setdiff(colnames(prevalence), "COUNTY_FIPS")
ZIP_FEATURES <- ZIP_FEATURES %>%
  select(-all_of(prev_columns))

# Count and print the number and percentage of missing values in the 'rate_prev_11/27/20' column
num_missing_rate_prev <- sum(is.na(ZIP_FEATURES$`prev_11/27/20`))
perc_missing_rate_prev <- mean(is.na(ZIP_FEATURES$`prev_11/27/20`)) * 100

cat("Number of missing values in rate_prev_11_27_20:", num_missing_rate_prev, "\n")
cat("Percentage of missing values in rate_prev_11_27_20:", perc_missing_rate_prev, "%\n")
```


Now let's merge ZIP_FEATURES to BICS to get a dataset that matches features of the ZIP code to the respondent
First, I'll read in the different waves seperately 
The two files have the same row count but different columns (225 in manual versus 62 in merged)
```{r}
if (file_type == "manual") {
  df_w2 <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/national_wave2_unweighted.rds')
  df_w3 <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/national_wave3_unweighted.rds')
  df_w4 <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/national_wave4_unweighted.rds')
  df_w5 <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/national_wave5_unweighted.rds')
  df_w6 <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/national_wave6_unweighted.rds')
  
  if("resp_employ" %in% names(df_w2)) {
  df_w2$resp_employ <- as.character(df_w2$resp_employ)
  }

  if("resp_employ" %in% names(df_w3)) {
  df_w3$resp_employ <- as.character(df_w3$resp_employ)
  }

  if("resp_employ" %in% names(df_w4)) {
  df_w4$resp_employ <- as.character(df_w4$resp_employ)
  }

  if("resp_employ" %in% names(df_w5)) {
  df_w5$resp_employ <- as.character(df_w5$resp_employ)
  }

  if("resp_employ" %in% names(df_w6)) {
  df_w6$resp_employ <- as.character(df_w6$resp_employ)
  }

  df_w2$resp_occupation <- as.character(df_w2$resp_occupation)
  df_w4$resp_occupation <- as.character(df_w4$resp_occupation)
  
  df <- bind_rows(df_w2, df_w3, df_w4, df_w5, df_w6)
  rm(df_w2, df_w3, df_w4, df_w5, df_w6)
  
  df_merged <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/df_all_waves.rds')

  df_merged <- df_merged %>% 
    filter(wave > 1)
  
  df$weight_pooled <- df_merged$weight_pooled
  df$weight_city <- df_merged$weight_city
  
  rm(df_merged)
  
} else if (file_type == "merged") {
  df <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/df_all_waves.rds')

  df <- df %>% 
    filter(wave > 1)
  
} else {
  print("change file_type to either manual or merged")
}

names(df) <- make.unique(names(df))

#deleting this for now and instead using city dummies
#df <- df %>% 
  #filter(city == "National")
```

```{r}
df$num_cc
```

Assigning dates to waves
```{r}
df <- df %>%
  mutate(data_collected_dates = case_when(
    wave == 2 ~ "June 18-24, '20",
    wave == 3 ~ "September 12-20, '20'",
    wave == 4 ~ "December 1-7, '20'",
    wave == 5 ~ "February 9-14, '21'",
    wave == 6 ~ "May 14-25, '21'",
    TRUE ~ as.character(NA)  # Equivalent to np.nan in Python
  ))

df$data_collected_dates <- factor(df$data_collected_dates, levels = c("June 18-24, '20", "September 12-20, '20'", "December 1-7, '20'","February 9-14, '21'","May 14-25, '21'"))  

print(table(df$data_collected_dates))
print(table(df$wave))
```
We're missing a "conservative" but we have two moderates?
```{r}
political_view_mapping <- c(
  'Extremely conservative' = 1,
  'Slightly conservative' = 2,
  'Moderate' = 3,
  'Middle of the road' = 4,
  'Slightly liberal' = 5,
  'Liberal' = 6,
  'Extremely liberal' = 7
)

df <- df %>%
  mutate(political_view_numeric = unlist(map(political_view, ~ political_view_mapping[.])))

print(table(df$political_view))
```
creating the education, race, and employed variables (merged file has no employment variable)
```{r}
if (file_type == "manual"){
df <- df %>%
  mutate(educ_group = case_when(
    resp_educ == "Less than high school degree" ~ "Less than high school",
    resp_educ %in% c("High school graduate (high school diploma or equivalent including GED)", 
                     "Some college but no degree", 
                     "Associate degree in college (2-year)") ~ "High school graduate",
    resp_educ %in% c("Bachelor's degree in college (4-year)", 
                     "Master's degree", 
                     "Doctoral degree", 
                     "Professional degree (JD, MD)") ~ "College graduate and above",
    TRUE ~ "Unknown"
  ))

df$educ_group <- as.factor(df$educ_group)

# Print the counts of unique values in educ_group
print(table(df$educ_group, useNA = "ifany"))

race_columns <- grep("^resp_race_", names(df), value = TRUE)
df[race_columns] <- lapply(df[race_columns], function(x) replace(x, is.na(x), 0))

df <- df %>%
  mutate(r_race = case_when(
    resp_race_1 == "White" ~ "White",
    resp_race_2 == "Black or African American" ~ "Black",
    resp_race_4 == "Asian" ~ "Asian",
    TRUE ~ "Other / Mixed"
  ))

# Print the counts of unique values in r_race
print(table(df$r_race, useNA = "ifany"))

employment_columns <- grep("^resp_employ_", names(df), value = TRUE)
df[employment_columns] <- lapply(df[employment_columns], function(x) replace(x, is.na(x), 0))

#this will not work because there are no employment variables in the merged file
df <- df %>%
  mutate(r_working = if_else(industry == "I don't work", "Not Working", "Working"))

# Print the counts of unique values in r_working
print(table(df$r_working, useNA = "ifany"))

} else if (file_type == "merged"){
  df$educ_group <- df$educ
  df$r_race <- df$w_race
  #there's no employment variable in the merged data
}
```
dependent binary variables
```{r}
df <- df %>%
  mutate(
    binary_concern = case_when(
      covid19_concern %in% c("Somewhat concerned", "Very concerned") ~ 1L,
      !is.na(covid19_concern) ~ 0L,
      TRUE ~ NA_integer_
    ),
    binary_concern_strong = case_when(
      covid19_concern == "Very concerned" ~ 1L,
      !is.na(covid19_concern) ~ 0L,
      TRUE ~ NA_integer_
    ),
    contact_reduction = case_when(
      covid19_f2fchange == "I have greatly reduced face-to-face interaction with others" ~ 1L,
      !is.na(covid19_f2fchange) ~ 0L,
      TRUE ~ NA_integer_
    )
  )
print(table(df$binary_concern_strong))
#there are some subset of people that will reduce contact but actually don't want to
#there are also some subet of people that won't reduce contact but actually want to
df <- df %>%
  mutate(binary_concern_strong = ifelse(covid19_reduceOK == "No, I would like to increase face-to-face interaction with others more than I am able to" & binary_concern_strong == 1, 0,
                                        ifelse(covid19_reduceOK == "No, I would like to reduce face-to-face interaction with others more than I am able to" & binary_concern_strong == 0, 1, binary_concern_strong)))

df <- df %>%
  mutate(strong_concern_inverse = ifelse(is.na(binary_concern_strong), NA,
                                        ifelse(binary_concern_strong == 1, 0,
                                               ifelse(binary_concern_strong == 0, 1, NA))))

print(table(df$binary_concern_strong))
```
manual merge = 27,937, already merged = 69,135
```{r}
if (file_type == "manual") {
  df_w2_nonhhalters <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/national_alters_nonhh_wave2_unweighted.rds')
  df_w3_nonhhalters <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/national_alters_nonhh_wave3_unweighted.rds')
  df_w4_nonhhalters <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/national_alters_nonhh_wave4_unweighted.rds')
  df_w5_nonhhalters <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/national_alters_nonhh_wave5_unweighted.rds')
  df_w6_nonhhalters <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/national_alters_nonhh_wave6_unweighted.rds')
  
  df_alters_combined <- rbind(df_w2_nonhhalters, df_w3_nonhhalters, df_w4_nonhhalters, df_w5_nonhhalters, df_w6_nonhhalters)
  rm(df_w2_nonhhalters, df_w3_nonhhalters, df_w4_nonhhalters, df_w5_nonhhalters, df_w6_nonhhalters)
} else if (file_type == "merged") {
  df_alters_combined <- readRDS('/Users/chrissoria/Documents/Research/BICS/bics-data/data/lucid-pipeline/df_alters_all_waves.rds')
}
```
Now I'll calculate the proportion of contacts carried out with masks (the merged file doesn't have this)
```{r}
if (file_type == "manual") {
df_alters_combined <- df_alters_combined %>%
  mutate(
    Mask_Used = as.integer(protection_1 == "Wear a face mask" |
                           protection_2 == "Wear a face mask" |
                           protection_3 == "Wear a face mask" |
                           protection_4 == "Wear a face mask"),
    Gloves_Used = as.integer(protection_1 == "Wear gloves" |
                             protection_2 == "Wear gloves" |
                             protection_3 == "Wear gloves" |
                             protection_4 == "Wear gloves"),
    Other_Protective_Equipment_Used = as.integer(protection_1 == "Wear other protective equipment" |
                                                  protection_2 == "Wear other protective equipment" |
                                                  protection_3 == "Wear other protective equipment" |
                                                  protection_4 == "Wear other protective equipment")
  )

df_aggregated <- df_alters_combined %>%
  group_by(rid) %>%
  summarise(
    Total_Masks_Used = sum(Mask_Used, na.rm = TRUE),
    Total_Gloves_Used = sum(Gloves_Used, na.rm = TRUE),
    Total_Other_Protective_Equipment_Used = sum(Other_Protective_Equipment_Used, na.rm = TRUE),
    Contacts = n()
  ) %>%
  mutate(
    Norm_Masks_Used = (Total_Masks_Used / Contacts) *100, #changing this to a percentage for ease of interpretation
    Norm_Gloves_Used = Total_Gloves_Used / Contacts,
    Norm_Other_Protective_Equipment_Used = Total_Other_Protective_Equipment_Used / Contacts,
    Non_Weighted_Safety_Index = (Norm_Masks_Used + Norm_Gloves_Used + Norm_Other_Protective_Equipment_Used) / 3,
    Weighted_Safety_Index = (Norm_Masks_Used + Norm_Gloves_Used + Norm_Other_Protective_Equipment_Used) / (3 * Contacts)
  )

df = left_join(df, df_aggregated, by = "rid")
} else if (file_type == "merged") {
  df = left_join(df, df_alters_combined, by = "rid")
  print("The merged file doesn't have information on mask usage")
}
```
Now let's merge in the ZIP features file
```{r}
ZIP_FEATURES <- fread('/Users/chrissoria/Documents/Research/BICS_Political_Polarization/data/ZIP_Features.csv', colClasses = c(ZIP = "character", CD = "character", COUNTY_FIPS = "character", STATE_FIPS = "character", CONGRESSPERSON_PARTY = "character"))
```

```{r}
unique_occupations <- unique(df$resp_occupation)

print(unique_occupations)

essential_categories <- c("Accommodation or food services", "Admin, support, waste management or remediation services", 
                          "Agriculture, Forestry, Fishing, and Hunting", "Forestry, fishing, hunting or agriculture support", 
                          "Manufacturing", "Construction", "Finance or insurance", "Mining", "Transportation or warehousing", 
                          "Utilities", "Government or Public Policy", "Community/Social Services", 
                          "Mining, Quarrying, Oil and Gas Extraction", "Transportation", "Warehousing/Logistics", 
                          "Educational services", "Professional, scientific or technical services", "Health care")

df <- df %>%
  mutate(essential = ifelse(resp_occupation %in% essential_categories, 1, 0))
```

getting df ready to merge to ZIP_FEATURES
merged file questions:
Why are there duplicates w_gender and gender, for example?
No nativity (US-born vs not), no resp_hhsize,lefthome_num, policy_sip (), health insurance, wave is called wave.x, city is called city.x
```{r}
if (file_type == "manual") {
df <- rename(df, ZIP = resp_zip)

names(df) <- make.unique(names(df))

# Trim whitespace from ZIP column
df$ZIP <- trimws(df$ZIP)

columns_to_keep <- c('ResponseId','StartDate','resp_yob','resp_sex','resp_hispanic','r_race','resp_nativity','ZIP',
                     'resp_hhsize','r_working','resp_occupation','lefthome_num','num_cc_nonhh',
                     'lefthome_where_1','lefthome_where_2','lefthome_where_3','lefthome_where_4','lefthome_where_10',
                     'lefthome_where_8','lefthome_where_9','lefthome_where_5','lefthome_where_11','lefthome_where_6',
                     'lefthome_where_7','inet_freq','covid19_familiar','covid19_concern',
                     'covid19_f2fchange','covid19_reduceOK','policy_sip','age','hhi','political_party','political_view',
                     'industry','health_insurance','interview_date','wave','agecat','city','covid19_vax','covid19_whynot_vax',
                     'Non_Weighted_Safety_Index','Weighted_Safety_Index','Norm_Masks_Used','educ_group','contact_reduction',
                     'binary_concern','binary_concern_strong','resp_educ','Contacts',
                     'data_collected_dates','weight_pooled','weight_city','num_cc','strong_concern_inverse','essential')

df <- df[columns_to_keep]

} else if (file_type == "merged") {
  df <- rename(df, ZIP = zip)
  df$ZIP <- trimws(df$ZIP)
  
  columns_to_keep <- c('ResponseId','age','gender','hispanic','w_race','ZIP',
                     'num_cc_nonhh','covid19_familiar','covid19_concern',
                     'covid19_f2fchange','covid19_reduceOK','hhi','political_party','political_view',
                     'wave.x','agecat','city.x','covid19_vax','covid19_whynot_vax',
                     'educ_group','contact_reduction',
                     'binary_concern','binary_concern_strong','educ','data_collected_dates')
  
  df <- df[columns_to_keep]
}
```
merging zip features to bics
```{r}
bics_zip_features <- left_join(df, ZIP_FEATURES, by = "ZIP")

bics_zip_features <- bics_zip_features %>% 
  distinct(ResponseId, .keep_all = TRUE)
```
How many matches do we get?
```{r}
bics_zip_features %>%
  summarise(
    Num_Missing_ZIP = sum(is.na(ZIP)),
    Pct_Missing_ZIP = mean(is.na(ZIP)) * 100,
    Pct_ZIP_5_Characters = mean(nchar(as.character(ZIP)) == 5) * 100,
    state_Missing = sum(is.na(bics_zip_features$USPS_STATE)),
    Pct_State_Missing = mean(!is.na(bics_zip_features$USPS_STATE)) * 100
    
  ) %>%
  print()
```
what about school closures?
About 9 percent of people have a missing school matched to their ZIP
But lots of missingess for the spring 2021
```{r}
bics_zip_features %>%
  summarise(
    Num_Missing_earlyfall2020 = sum(is.na(earlyfall2020_school_closures)),
    Pct_Missing_earlyfall2020 = mean(is.na(earlyfall2020_school_closures)) * 100,
    Num_Missing_fall2020 = sum(is.na(fall2020_school_closures)),
    Pct_Missing_fall2020 = mean(is.na(fall2020_school_closures)) * 100,
    Num_Missing_spring2021 = sum(is.na(spring2021_school_closures)),
    Pct_Missing_spring2021 = mean(is.na(spring2021_school_closures)) * 100
  ) %>%
  print()
```
And county-level mask mandates? Only .5% of people are missing this information.
```{r}
bics_zip_features %>%
  summarise(
    Num_Missing_maskmandates = sum(is.na(county_conditions)),
    Pct_Missing_maskmandates = mean(is.na(county_conditions)) * 100
  ) %>%
  print()
```
Let's clean up the mask mandates variable
```{r}
strict_mandates <- c("all public places","Face coverings are required to be worn over the mouth and nose of all county residents while inside or on grounds of all government buildings, private businesses, organizations or venues open to the public.","in public places","indoor and outdoor places","Indoor, outdoor, public places and public transportations","most public places", "public places where 6 feet of social distance cannot be maintained","All employees and members of the general public","Face masks are now required when exercising, even outdoors.","Require every person in the city to wear a face covering when inside a commercial business or other structure, building or space open to the public. Masks also will be required in an outdoor space when staying 6 feet apart for social distancing from someone not of the same household is difficult.  (passed by Tulsa City Council)","State leaders targeted Butler County as a red-alert area Tuesday, requiring all residents to wear protective masks in public to slow the recent local spike in coronavirus cases.","All employees and members of the general public all public places","public places where 6 feet of social distance cannot be maintained","public places and where social distancing impossible")

less_strict_mandates <- c("Require masks in indoor public places throughout Oklahoma City. (a mask mandate passed by Oklahoma City Coucil; however, a similar mandate denied by the county commissioners)","All customers in retail facilities, most forward facing facilities and if you are indoors or outdoors and cannot be 6ft away from another person, you must wear a mask.","And in 7/15 people were asked to wear masks when entering the government buildings.","Anderson City Council approves mask ordinance in 7/23","Anyone over age 2 who can medically tolerate a face covering when in a public place and unable to maintain social distancing.","counties with more than 20 coronavirus cases to wear a face covering over the nose and mouth while in a business or other building open to the public, as well as outdoor public spaces, whenever social distancing is not possible.","Employees working with the public","following the state order","in all types of businesses","indoor commercial businesses and town offices in the town of Boone","indoor places and indoor businesses, workers need to wear one if outdoor not social distancing impossible","Indoor public spaces and at any organized outdoor activity where social distancing cannot be maintained, including markets, weddings and parties","inside public spaces except when eating or drinking","retail only","The Clinton County Board of Supervisors voted in favor of requiring face masks or face shields to be worn in all Clinton County buildings starting Tuesday, July 27.","applies to any indoor or confined public setting where a person will be within six feet of another individual, who does not share the same household(5/27). Follow the state order(7/10).","Indoor public spaces and at any organized outdoor activity where social distancing cannot be maintained, including markets, weddings and parties","The directive applies to any indoor or confined public setting where a person will be within six feet of another individual, who does not share the same household(5/27). Follow the state order(7/10).","Face masks no longer mandatory for Bell County businesses started from 6_29, but Kentucky ordered to wear masks started from 7_10_2020")

unclear_or_unspecified_mandates <- c("unspecified", "unclear","other")

no_mandates <- c("Local officials generally agreed with the idea behind the mandate, but reactions differed on the execution of the order as well as if it was necessary at all. The public reaction locally also varied wildly, ranging from angered to very pleased based on social media input.","no mandate","opt out","opts out of the provisions of Executive Order No. 20-52 requiring masks or other face coverings in public but recommend the use of face coverings","Despite not passing a city-wide masking requirement, Nixa city leaders have recommended that people follow the Centers for Disease Control and Prevention guidelines about wearing face coverings in public.")

bics_zip_features <- bics_zip_features %>% 
  mutate(county_conditions = trimws(county_conditions)) %>%
  mutate(county_mask_mandate = case_when(
    county_conditions %in% strict_mandates ~ "Strict",
    county_conditions %in% less_strict_mandates ~ "Less Strict",
    county_conditions %in% unclear_or_unspecified_mandates ~ "Unspecified or Unclear Mandate",
    county_conditions %in% no_mandates ~ "No Mandate",
    TRUE ~ NA_character_
  )) %>%
  mutate(county_mask_mandate = factor(county_mask_mandate, levels = c("Strict", "Less Strict", "Unspecified or Unclear Mandate", "No Mandate")))


print(table(bics_zip_features$county_conditions, useNA = "always"))
print(table(bics_zip_features$county_mask_mandate, useNA = "always"))
```

```{r}
bics_zip_features <- bics_zip_features %>%
  mutate(CONGRESSPERSON_PARTY = ifelse(CONGRESSPERSON_PARTY == "", NA_character_,
                     ifelse(CONGRESSPERSON_PARTY == " ", NA_character_, CONGRESSPERSON_PARTY)))

bics_zip_features <- bics_zip_features %>%
  mutate(political_party_to_CD = ifelse(!is.na(CONGRESSPERSON_PARTY), 
                                        paste(political_party, "in", CONGRESSPERSON_PARTY, "led CD"), 
                                        NA_character_))

bics_zip_features <- bics_zip_features %>%
  mutate(political_party_to_CD = ifelse(political_party == "Prefer not to answer", NA_character_, political_party_to_CD))

bics_zip_features <- bics_zip_features %>%
  mutate(political_party_to_CD = ifelse(is.na(political_party_to_CD), "unknown party or CD", political_party_to_CD))

print(table(bics_zip_features$CONGRESSPERSON_PARTY, useNA = "always"))
print(table(bics_zip_features$political_party_to_CD, useNA = "always"))
```
creating a variable that accounts for ideology
```{r}
bics_zip_features <- bics_zip_features %>%
  mutate(political_view_to_CD = ifelse(!is.na(CONGRESSPERSON_PARTY), 
                                        paste(political_view, "in", CONGRESSPERSON_PARTY, "led CD"), 
                                        NA_character_))

print(table(bics_zip_features$CONGRESSPERSON_PARTY, useNA = "always"))
print(table(bics_zip_features$political_view_to_CD, useNA = "always"))
```


generating binary variables for whether someone is in an opposing party CD (to observe potential general effects), vaccinated binaries, and Independent
```{r}
bics_zip_features <- bics_zip_features %>%
  mutate(
    In_Opposing_Party_CD = as.integer(political_party != CONGRESSPERSON_PARTY & !is.na(CONGRESSPERSON_PARTY) & political_party != "Prefer not to answer"),
    Independent = as.integer(political_party == "Independent"),
    Vaccinated = case_when(
      covid19_vax == "Yes, I have received at least one dose of a vaccine" ~ 1L,
      covid19_vax == "No, I have not" ~ 0L,
      TRUE ~ NA_integer_
    )
  )

# Print the counts of unique values in In_Opposing_Party_CD column
print(table(bics_zip_features$In_Opposing_Party_CD, useNA = "ifany"))

# Print the counts of unique values in political_party column
print(table(bics_zip_features$political_party, useNA = "ifany"))

# Print the counts of unique values in CONGRESSPERSON_PARTY column
print(table(bics_zip_features$CONGRESSPERSON_PARTY, useNA = "ifany"))

```
Next, I'm creating categorical variables based on percentage of the county voting a certain way
```{r}
bics_zip_features <- bics_zip_features %>%
  mutate(
    Categorical_Trump_County_Share = case_when(
      trump_percentage_won < 33.1 ~ "Less Than a Third",
      trump_percentage_won < 66.1 ~ "Less than Two Thirds Greater Than One Third",
      trump_percentage_won >= 66.1 ~ "Greater than or Equal to Two thirds",
      TRUE ~ NA_character_
    ),
    Categorical_Biden_County_Share = case_when(
      biden_percentage_won < 33.1 ~ "Less Than a Third",
      biden_percentage_won < 66.1 ~ "Less than Two Thirds Greater Than One Third",
      biden_percentage_won >= 66.1 ~ "Greater than or Equal to Two thirds",
      TRUE ~ NA_character_
    ),
    Percentage_Trump_Greater_Than_Two_Thirds = case_when(
      trump_percentage_won >= 66.1 ~ 1L,
      trump_percentage_won < 66.1 ~ 0L,
      TRUE ~ NA_integer_
    ),
    Percentage_Trump_Less_Than_Two_Thirds = case_when(
      trump_percentage_won < 66.1 ~ 1L,
      trump_percentage_won <= 66.1 ~ 0L,
      TRUE ~ NA_integer_
    ),
    Percentage_Biden_Greater_Than_Two_Thirds = case_when(
      biden_percentage_won >= 66.1 ~ 1L,
      biden_percentage_won < 66.1 ~ 0L,
      TRUE ~ NA_integer_
    ),
    Percentage_Biden_Less_Than_Two_Thirds = case_when(
      biden_percentage_won < 66.1 ~ 1L,
      biden_percentage_won <= 66.1 ~ 0L,
      TRUE ~ NA_integer_
    )
  )
print(table(bics_zip_features$Categorical_Trump_County_Share))
print(table(bics_zip_features$Categorical_Biden_County_Share))
```
And also based on CD
```{r}
bics_zip_features <- bics_zip_features %>%
  mutate(
    Categorical_Repub_CD_County_Share = case_when(
      CD_PERCENT_REPUBLICAN < 0.25 ~ "Less Than a Quarter",
      CD_PERCENT_REPUBLICAN < 0.50 & CD_PERCENT_REPUBLICAN > 0.25 ~ "Two Quarters",
      CD_PERCENT_REPUBLICAN < 0.75 & CD_PERCENT_REPUBLICAN > 0.50 ~ "Three Quarters",
      CD_PERCENT_REPUBLICAN >= 0.75 ~ "Greater than or Equal to Three Quarters",
      TRUE ~ NA_character_
    ),
    Categorical_Dem_CD_County_Share = case_when(
      CD_PERCENT_DEMOCRAT < 0.25 ~ "Less Than a Quarter",
      CD_PERCENT_DEMOCRAT < 0.50 ~ "Two Quarters",
      CD_PERCENT_DEMOCRAT < 0.75 ~ "Three Quarters",
      CD_PERCENT_DEMOCRAT >= 0.75 ~ "Greater than or Equal to Three Quarters",
      TRUE ~ NA_character_
    )
  )

# Print the counts of unique values in Categorical_Repub_CD_County_Share column
print(table(bics_zip_features$Categorical_Repub_CD_County_Share, useNA = "ifany"))
```
Next, let's add a variable that treats urban and rural as a binary.
1 = Metro - Counties in metro areas of 1 million population or more
2 = Metro - Counties in metro areas of 250,000 to 1 million population \
3 = Metro - Counties in metro areas of fewer than 250,000 population
4 = Nonmetro - Urban population of 20,000 or more, adjacent to a metro area
5 = Nonmetro - Urban population of 20,000 or more, not adjacent to a metro area
6 = Nonmetro - Urban population of 2,500 to 19,999, adjacent to a metro area
7 = Nonmetro - Urban population of 2,500 to 19,999, not adjacent to a metro area
8 = Nonmetro - Completely rural or less than 2,500 urban population, adjacent to a metro area
9 = Nonmetro - Completely rural or less than 2,500 urban population, not adjacent to a metro area
```{r}
bics_zip_features <- bics_zip_features %>%
  mutate(
    COUNT_RUCC_CAT = case_when(
      COUNTY_RUCC_2013 %in% c(1, 2) ~ "Urban Metro",
      COUNTY_RUCC_2013 %in% 3:7 ~ "Urban Nonmetro",
      COUNTY_RUCC_2013 %in% c(8, 9) ~ "Rural",
      TRUE ~ NA_character_
    ),
    COUNT_RUCC_BINARY = case_when(
      COUNTY_RUCC_2013 %in% 1:7 ~ "Urban",
      COUNTY_RUCC_2013 %in% c(8, 9) ~ "Rural",
      TRUE ~ NA_character_
    )
  )
```

```{r}
rm(df, df_alters_combined, ZIP_FEATURES, columns_to_keep, employment_columns, political_view_mapping, race_columns, df_aggregated)
```
Making the column names unique (in case i merged in duplicate columns somehow)
```{r}
names(bics_zip_features) <- make.unique(names(bics_zip_features))
```
Making the percent of congressional district that voted a certain way be a categorical variable and setting the reference category
```{r}
bics_zip_features$Categorical_Dem_CD_County_Share <- factor(
  bics_zip_features$Categorical_Dem_CD_County_Share,
  levels = c("Greater than or Equal to Three Quarters", "Three Quarters", "Two Quarters", "Less Than a Quarter"),
  ordered = FALSE
)

bics_zip_features$Categorical_Repub_CD_County_Share <- factor(
  bics_zip_features$Categorical_Repub_CD_County_Share,
  levels = c("Greater than or Equal to Three Quarters", "Three Quarters", "Two Quarters", "Less Than a Quarter"),
  ordered = FALSE
)
```
We might want to control for city samples, so let's pull out some dummies
```{r}
bics_zip_features$city <- as.factor(bics_zip_features$city) # Ensure 'city' is a factor
dummy_city <- model.matrix(~ city - 1, data = bics_zip_features) # Create dummy variables
bics_zip_features <- cbind(bics_zip_features, dummy_city)
rm(dummy_city)
```
The income variable is a bit too high-resolution to use in a model as a categorical variable
Pulling out party-specific data sets 
```{r}
bics_zip_features <- bics_zip_features %>%
  mutate(Income_Category = case_when(
    hhi %in% c("Less than $5,000", "$5,000 to $9,999", "$10,000 to $14,999", "$15,000 to $19,999", "$20,000 to $24,999", "$25,000 to $29,999") ~ "Less than $30,000",
    hhi %in% c("$30,000 to $34,999", "$35,000 to $39,999", "$40,000 to $44,999", "$45,000 to $49,999", "$50,000 to $54,999", "$55,000 to $59,999") ~ "$30,000 to $59,999",
    hhi %in% c("$60,000 to $64,999", "$65,000 to $69,999", "$70,000 to $74,999", "$75,000 to $79,999", "$80,000 to $84,999", "$85,000 to $89,999") ~ "$60,000 to $89,999",
    hhi %in% c("$90,000 to $94,999", "$95,000 to $99,999", "$100,000 to $124,999", "$125,000 to $149,999", "$150,000 to $174,999", "$175,000 to $199,999") ~ "$90,000 to $199,999",
    hhi %in% c("$200,000 to $249,999", "$250,000 and above") ~ "$200,000 and above",
    TRUE ~ "Prefer not to answer"
  ))

bics_zip_features$Income_Category <- as.factor(bics_zip_features$Income_Category)
bics_zip_features$Income_Category <- factor(bics_zip_features$Income_Category, levels = c("Less than $30,000", "$30,000 to $59,999", "$60,000 to $89,999","$90,000 to $199,999","200,000 to $249,999", "Prefer not to answer"))
```
setting sex to factor, creating dummies for male, race-white, working
```{r}
bics_zip_features$resp_sex <- as.factor(bics_zip_features$resp_sex)
bics_zip_features$resp_sex <- factor(bics_zip_features$resp_sex, levels = c("Male","Female","Other"))

#male
bics_zip_features <- bics_zip_features %>% 
  mutate(male = case_when(
    is.na(resp_sex) ~ NA_real_,
    resp_sex == "Male" ~ 1,
    TRUE ~ 0
  ))
#white
bics_zip_features <- bics_zip_features %>% 
  mutate(white = case_when(
    is.na(r_race) ~ NA_real_,
    r_race == "White" ~ 1,
    TRUE ~ 0
  ))
#hispanic
bics_zip_features <- bics_zip_features %>% 
  mutate(r_hispanic_num = case_when(
    is.na(resp_hispanic) ~ NA_real_,
    resp_hispanic == "Yes" ~ 1,
    TRUE ~ 0
  ))
#working
bics_zip_features <- bics_zip_features %>% 
  mutate(r_working_num = case_when(
    is.na(r_working) ~ NA_real_,
    r_working == "Working" ~ 1,
    TRUE ~ 0
  ))
#college
bics_zip_features <- bics_zip_features %>% 
  mutate(r_college_grad = case_when(
    is.na(educ_group) ~ NA_real_,
    educ_group == "College graduate and above" ~ 1,
    TRUE ~ 0
  ))
#urban
bics_zip_features <- bics_zip_features %>% 
  mutate(c_urban = case_when(
    is.na(COUNT_RUCC_BINARY) ~ NA_real_,
    COUNT_RUCC_BINARY == "Urban" ~ 1,
    TRUE ~ 0
  ))
#over 65 or below
bics_zip_features <- bics_zip_features %>%
  mutate(young = case_when(
    is.na(age) ~ NA_real_,  # Preserve NA values by explicitly handling them
    age < 64 ~ 1,
    TRUE ~ 0
  ))

#republican
bics_zip_features <- bics_zip_features %>%
  mutate(republican = case_when(
    is.na(political_party) ~ NA_real_,  # Preserve NA values by explicitly handling them
    political_party == "Republican" ~ 1,
    TRUE ~ 0
  ))

#democrat
bics_zip_features <- bics_zip_features %>%
  mutate(democrat = case_when(
    is.na(political_party) ~ NA_real_,  # Preserve NA values by explicitly handling them
    political_party == "Democrat" ~ 1,
    TRUE ~ 0
  ))

#republican district
bics_zip_features <- bics_zip_features %>%
  mutate(republican_district = case_when(
    is.na(CONGRESSPERSON_PARTY) ~ NA_real_,  # Preserve NA values by explicitly handling them
    CONGRESSPERSON_PARTY == "Republican" ~ 1,
    TRUE ~ 0
  ))

#democrat district
bics_zip_features <- bics_zip_features %>%
  mutate(democrat_district = case_when(
    is.na(CONGRESSPERSON_PARTY) ~ NA_real_,  # Preserve NA values by explicitly handling them
    CONGRESSPERSON_PARTY == "Democratic" ~ 1,
    TRUE ~ 0
  ))

#metro
bics_zip_features <- bics_zip_features %>%
  mutate(Metro = case_when(
    COUNTY_RUCC_2013 < 3 ~ 1,
    COUNTY_RUCC_2013 > 3 ~ 0,
    TRUE ~ NA_real_ 
  ))
```
Now, it might be the case that political orientation is more important that political party. For example, Republicans in certain regions might actually just be less conservative, and therefore it might not be the party affiliation that matters but political orientation
```{r}
bics_zip_features <- bics_zip_features %>% 
  mutate(
    extreme_liberal = case_when(
      political_view == "Extremely liberal" ~ 1,
      is.na(political_view) ~ NA_real_,
      TRUE ~ 0
    ),
    extreme_conservative = case_when(
      political_view == "Extremely conservative" ~ 1,
      is.na(political_view) ~ NA_real_,
      TRUE ~ 0
    ),
    moderate = case_when(
      political_view == "Moderate" ~ 1,
      is.na(political_view) ~ NA_real_,
      TRUE ~ 0
    )
  )

bics_zip_features <- bics_zip_features %>% 
  mutate(
    liberal = case_when(
      political_view %in% c("Extremely liberal","Liberal","Slightly Liberal") ~ 1,
      is.na(political_view) ~ NA_real_,
      TRUE ~ 0
    ),
    conservative = case_when(
      political_view %in% c("Extremely conservative","Middle of the road", "Slightly conservative") ~ 1,
      is.na(political_view) ~ NA_real_,
      TRUE ~ 0
    )
  )

bics_zip_features$political_view <- as.factor(bics_zip_features$political_view)
```
These variables should be logged so that they fit better with a linear model
using log1p for handling zeros
```{r}

bics_zip_features <- bics_zip_features %>% 
  mutate(log_rate_prev_12_01_20 = log(`rate_prev_12/1/20`+ 1),
         log_rate_prev_6_14_20 = log(`rate_prev_6/14/20`+1),
         log_rate_prev_9_9_20 = log(`rate_prev_9/9/20`+1),
         log_rate_prev_2_4_21 = log(`rate_prev_2/4/21`+1),
         log_rate_prev_5_12_21 = log(`rate_prev_5/12/21`+1))  #

```
Lastly, we decided that instead of cumilative cases we want the new cases from the week prior. 
First, I'll read in the incidence data.
Second, I'll create variables to match on.
Third, I'll merge. 

```{r}
inc_w2 <- read_csv('../data/wave_2_incidence.csv', col_types = cols(FIPS = col_character())) %>% 
  rename(COUNTY_FIPS = FIPS,
         current_week_incidence = wave_2_incidence,
         prev_week_incidence = wave_2_inc_week_lag)

inc_w3 <- read_csv('../data/wave_3_incidence.csv', col_types = cols(FIPS = col_character())) %>% 
  rename(COUNTY_FIPS = FIPS,
         current_week_incidence = wave_3_incidence,
         prev_week_incidence = wave_3_inc_week_lag)

inc_w4 <- read_csv('../data/wave_4_incidence.csv', col_types = cols(FIPS = col_character())) %>% 
  rename(COUNTY_FIPS = FIPS,
         current_week_incidence = wave_4_incidence,
         prev_week_incidence = wave_4_inc_week_lag)

inc_w5 <- read_csv('../data/wave_5_incidence.csv', col_types = cols(FIPS = col_character())) %>% 
  rename(COUNTY_FIPS = FIPS,
         current_week_incidence = wave_5_incidence,
         prev_week_incidence = wave_5_inc_week_lag)

inc_w6 <- read_csv('../data/wave_6_incidence.csv') %>%
  rename(COUNTY_FIPS = FIPS,
         current_week_incidence = wave_6_incidence,
         prev_week_incidence = wave_6_inc_week_lag)

inc_w6 <- inc_w6 %>%
  mutate(COUNTY_FIPS = floor(COUNTY_FIPS))

inc_w6 <- inc_w6 %>%
  mutate(COUNTY_FIPS = as.character(COUNTY_FIPS))

waves <- list()

for(i in 2:6) {
  waves[[paste("wave", i, sep = "")]] <- bics_zip_features %>%
    filter(wave == i)
}

wave2 <- waves[["wave2"]]
wave3 <- waves[["wave3"]]
wave4 <- waves[["wave4"]]
wave5 <- waves[["wave5"]]
wave6 <- waves[["wave6"]]

wave2 <- wave2 %>%
  mutate(COUNTY_FIPS = str_pad(COUNTY_FIPS, width = 5, pad = "0", side = "left"))
wave3 <- wave3 %>%
  mutate(COUNTY_FIPS = str_pad(COUNTY_FIPS, width = 5, pad = "0", side = "left"))
wave4 <- wave4 %>%
  mutate(COUNTY_FIPS = str_pad(COUNTY_FIPS, width = 5, pad = "0", side = "left"))
wave5 <- wave5 %>%
  mutate(COUNTY_FIPS = str_pad(COUNTY_FIPS, width = 5, pad = "0", side = "left"))
wave6 <- wave6 %>%
  mutate(COUNTY_FIPS = str_pad(COUNTY_FIPS, width = 5, pad = "0", side = "left"))

inc_w2 <- inc_w2 %>%
  mutate(COUNTY_FIPS = str_pad(COUNTY_FIPS, width = 5, pad = "0", side = "left"))
inc_w3 <- inc_w3 %>%
  mutate(COUNTY_FIPS = str_pad(COUNTY_FIPS, width = 5, pad = "0", side = "left"))
inc_w4 <- inc_w4 %>%
  mutate(COUNTY_FIPS = str_pad(COUNTY_FIPS, width = 5, pad = "0", side = "left"))
inc_w5 <- inc_w5 %>%
  mutate(COUNTY_FIPS = str_pad(COUNTY_FIPS, width = 5, pad = "0", side = "left"))
inc_w6 <- inc_w6 %>%
  mutate(COUNTY_FIPS = str_pad(COUNTY_FIPS, width = 5, pad = "0", side = "left"))


wave2 <- left_join(wave2, inc_w2, by = "COUNTY_FIPS") %>%
  mutate(current_week_inc_rate = (current_week_incidence / POPESTIMATE2020) * 100000,
         prev_week_inc_rate = (prev_week_incidence / POPESTIMATE2020) * 100000)

wave3 <- left_join(wave3, inc_w3, by = "COUNTY_FIPS") %>%
  mutate(current_week_inc_rate = (current_week_incidence / POPESTIMATE2020) * 100000,
         prev_week_inc_rate = (prev_week_incidence / POPESTIMATE2020) * 100000)

wave4 <- left_join(wave4, inc_w4, by = "COUNTY_FIPS") %>%
  mutate(current_week_inc_rate = (current_week_incidence / POPESTIMATE2020) * 100000,
         prev_week_inc_rate = (prev_week_incidence / POPESTIMATE2020) * 100000)
print(range(wave4$current_week_inc_rate))

wave5 <- left_join(wave5, inc_w5, by = "COUNTY_FIPS") %>%
  mutate(current_week_inc_rate = (current_week_incidence / POPESTIMATE2021) * 100000,
         prev_week_inc_rate = (prev_week_incidence / POPESTIMATE2021) * 100000)

wave6 <- left_join(wave6, inc_w6, by = "COUNTY_FIPS") %>%
  mutate(current_week_inc_rate = (current_week_incidence / POPESTIMATE2021) * 100000,
         prev_week_inc_rate = (prev_week_incidence / POPESTIMATE2021) * 100000)

bics_zip_features <- bind_rows(wave2, wave3, wave4, wave5, wave6)

bics_zip_features <- bics_zip_features %>%
  distinct(ResponseId, .keep_all = TRUE)

print(sum(is.na(bics_zip_features$COUNTY_FIPS)))
print(sum(is.na(bics_zip_features$prev_week_inc_rate)))
```
```{r}
rm(waves, wave2,wave3,wave4,wave5,wave6,inc_w2,inc_w3,inc_w4,inc_w5,inc_w6)
rm(df_w2,df_w3,df_w4,df_w5,df_w6)
```
After looking at the distribution of the incidence rate I noticed it's very right skewed.
```{r}
hist(bics_zip_features$prev_week_inc_rate, main="Histogram of Previous Week Incidence Rate",
     xlab="Previous Week Incidence Rate", breaks=50, col="grey")
```
I will log this to make it a more normal distribution (a better fit for the linear models we'll be using in the early stage of work). log1p for handling zeros. 

There were about 200 cases where the incidence rate is negative. Probably due to an error in reporting OR something else. For now, I'll drop all of those cases. 

Afterwards, the distribution is still slightly skwewed, but much less so. 
```{r}
bics_zip_features <- bics_zip_features %>%
  filter(current_week_inc_rate > 0)

bics_zip_features <- bics_zip_features %>% 
  mutate(
    log_current_week_inc_rate = log(current_week_inc_rate + 1),
    log_prev_week_inc_rate = log(prev_week_inc_rate + 1)
  )

hist(bics_zip_features$log_current_week_inc_rate, main="Histogram of Logged Previous Week Incidence Rate",
     xlab="Logged Previous Week Incidence Rate", breaks=50, col="grey")
```
```{r}
bics_zip_features$wave <- as.factor(bics_zip_features$wave)
bics_zip_features$Vaccinated <- bics_zip_features$Vaccinated*100
```

```{r}
RD_df <- bics_zip_features[bics_zip_features$political_party %in% c("Republican", "Democrat"), ]
D_df <- subset(bics_zip_features, political_party == "Democrat")
R_df <- subset(bics_zip_features, political_party == "Republican")
I_df <- subset(bics_zip_features, political_party == "Independent")

D_df <- subset(D_df, !is.na(D_df$Categorical_Repub_CD_County_Share))
R_df <- subset(R_df, !is.na(R_df$Categorical_Dem_CD_County_Share))

national_df <- subset(bics_zip_features, city == "National")
```
```{r}
write_csv(bics_zip_features, "../data/BICS_ZIP_Features.csv")
write_csv(RD_df, "../data/RD_BICS_ZIP_Features.csv")
write_csv(D_df, "../data/D_BICS_ZIP_Features.csv")
write_csv(R_df, "../data/R_BICS_ZIP_Features.csv")
```
building the df for weekly case counts
```{r}
library(lubridate)

dates <- c("June 18-24, '20", "September 12-20, '20", "December 1-7, '20", "February 9-14, '21", "May 14-25, '21")
dates_formal <- c("2020-06-18", "2020-09-12", "2020-12-01", "2021-02-09", "2021-05-14")
dates_formal <-ymd(dates_formal)

weekly_cases <- c(693,772,2351,2192,545)

cases_df <- data.frame(wave = dates, weekly_cases = weekly_cases, formal_dates = dates_formal)

cases_df$wave <- factor(cases_df$wave, levels = cases_df$wave)

rm(dates,weekly_cases)
```

```{r}
rm(file_type, no_mandates,strict_mandates,less_strict_mandates, unclear_or_unspecified_mandates, i, dates_formal)
```
