==> code/02-descriptives.Rmd <==
---
title: "01-descriptives"
author: "Christopher Soria"
date: "2024-03-08"
output: html_document
---

```{r}
library(tidyverse)
library(dplyr)
library(emmeans)
library(MASS)
library(emmeans)
```
Overall Findings and questions:
1. Working is correlated with the dependent variables but not independent variables of interest (political party). Do we still control for this?
2. Republicans within democratic districts look to be more likely to be employed than republicans in republican districts. The opposite is true for Democrats (probably a function of being in locations that are easier to get jobs in). Does it make sense to control for employment only in the models where I'm comparing same party identifcation in these different settings? EMPLOY ALL
3. Some of these differences probably change through the pandemic, would it be worth to look at how these averages differ throughout the pandemic?
4. Republicans in highly democratic congressional districts are less likely to be extremely conservative. Does this mean that it's important to control extreme conservative and extreme liberal in models where I'm trying to get at the difference between the two groups (those in opposing districts versus those in their own)?
5. Compared to Republicans, Democrats are much less likely to live in an opposing party district. Could this be a result of how sample was collected?
5. There shouldn't be differences in these outcomes by wave. For example, in wave 2 it should not be more or less likely that republicans versus democrats have smaller household sizes, etc. What are the implications on this? Does it even make sense to control for wave when we're conducting a pooled analysis (with the interpretation being something like, regardless of which wave of data this coming for, democrats are more likely to XYZ. But does including wave dilute any of the effects of interest?) YES CONTROL FOR HOUSEHOLD SIZE
6. Does it still make sense to use the pooled weights when I'm subsetting for political party affiliation?


Between Party Differences
High level findings:
- It probably doesn't need to be that we control for whether the respondent is working when comparing republican versus democrat (type of employment might be more important)

Detailed findings:
- Democrats have slightly smaller household sizes (maybe not statistically significant)
- Democrats are more likely to be Hispanic
- Democrats are less likely to be White
- Independents are less likely to be employed (maybe not statistically significant)
- Democrats are more likely to be in densely urbanized areas than both independents and republicans
- Democrats are more likely to have at least a college education than both independents and republicans
- Independents are the least likely to get a college degree
- Republicans are more likely to be male
- Republicans are older than both democrats and independents
```{r}
filtered_df <- bics_zip_features %>% 
  filter(political_party != "Prefer not to answer") %>% 
  filter(!is.na(political_party))

average_by_party <- filtered_df %>%
  group_by(political_party) %>%
  summarise(
    average_age = weighted.mean(age, w = weight_pooled, na.rm = TRUE),
    average_male = weighted.mean(male, w = weight_pooled, na.rm = TRUE),
    average_college = weighted.mean(r_college_grad, w = weight_pooled, na.rm = TRUE),
    average_employed = weighted.mean(r_working_num, w = weight_pooled, na.rm = TRUE),
    average_white = weighted.mean(white, w = weight_pooled, na.rm = TRUE),
    average_hispanic = weighted.mean(r_hispanic_num, w = weight_pooled, na.rm = TRUE),
    average_household_size = weighted.mean(resp_hhsize, w = weight_pooled, na.rm = TRUE),
    average_metro = weighted.mean(Metro, w = weight_pooled, na.rm = TRUE),
    num_total_cc = weighted.mean(num_cc, w = weight_pooled, na.rm = TRUE),
    num_non_houshold_cc = weighted.mean(num_cc_nonhh, w = weight_pooled, na.rm = TRUE),
    num_cc = weighted.mean(num_cc, w = weight_pooled, na.rm = TRUE),
    percent_mask = weighted.mean(Norm_Masks_Used, w = weight_pooled, na.rm = TRUE),
    vaccinated = weighted.mean(Vaccinated, w = weight_pooled, na.rm = TRUE),
    concern = weighted.mean(binary_concern_strong, w = weight_pooled, na.rm = TRUE)
  )

average_by_party

write_csv(average_by_party, "../data/party_average_differences.csv")
```
Do average contact and mask usage rates differ much over time?

```{r}
filtered_df <- bics_zip_features %>% 
  filter(political_party != "Prefer not to answer") %>% 
  filter(!is.na(political_party))

average_by_party_wave <- filtered_df %>%
  group_by(political_party, wave) %>%
  summarise(
    contacts = weighted.mean(num_cc_nonhh, w = weight_pooled, na.rm = TRUE),
    mask_usage = weighted.mean(Norm_Masks_Used, w = weight_pooled, na.rm = TRUE),
    concern = weighted.mean(binary_concern_strong, w = weight_pooled, na.rm = TRUE)
  )

average_by_party_wave

average_by_party_wave <- as.data.frame(average_by_party_wave)

rm(average_by_party_wave)
```


What about concern about the virus?

High level findings:
-It's important to control for all of these

Detailed findings: 
- (surprisingly) the younger are more likely to be concerned
- (surprisingly) male are more likely to say they are extremely concerned
- the college educated are more likely to say they're concerned
- the concerned are more likely to be non-white
- the concerned are more likely to be hispanic
- the concerned come from larger household family sizes
- the concerend are more likely to come from more densely urbanized environments
```{r}
average_by_concern <- filtered_df %>%
  group_by(binary_concern_strong) %>%
  summarise(
    average_age = mean(age, na.rm = TRUE),
    average_male = mean(male, na.rm = TRUE),
    average_college = mean(r_college_grad, na.rm = TRUE),
    average_employed = mean(r_working_num, na.rm = TRUE),
    average_white = mean(white, na.rm = TRUE),
    average_hispanic = mean(r_hispanic_num, na.rm = TRUE),
    average_household_size = mean(resp_hhsize, na.rm = TRUE),
    average_urban = mean(COUNTY_RUCC_2013, na.rm = TRUE)
  )

average_by_concern
```
Now let's look by vaccination status
High level findings
-Differences in all of these for vaccination status, especially for college educated, not much difference between hispanic and non-hispanics

Detailed findings: 
- people who are vaccinated are older
- people who are vaccinated are more likely to be male
- people who are vaccinated are WAY more likely to be college educated
- people who are vaccinated are more likely to be employed
- people who are vaccinated are more likely to white
- no difference for Hispanic
- people who are vaccinated are coming from smaller household sizes
- vaccinated people are more likely to come from densely urbanized counties
```{r}
average_by_vaccine <- filtered_df %>%
  group_by(Vaccinated) %>%
  filter(!is.na(Vaccinated)) %>% 
  summarise(
    average_age = mean(age, na.rm = TRUE),
    average_male = mean(male, na.rm = TRUE),
    average_college = mean(r_college_grad, na.rm = TRUE),
    average_employed = mean(r_working_num, na.rm = TRUE),
    average_white = mean(white, na.rm = TRUE),
    average_hispanic = mean(r_hispanic_num, na.rm = TRUE),
    average_household_size = mean(resp_hhsize, na.rm = TRUE),
    average_urban = mean(COUNTY_RUCC_2013, na.rm = TRUE)
  )

average_by_vaccine
```



What about reported non-household contacts? Breaking out into above and below average

High level findings:
- Even though some of these are only slightly different between groups, all of these categories differ in how many non-household contacts they report

Detailed findings:
- (surprisingly) people who report above average non-household contacts are younger
- people who report above average non-household contacts are slightly more likely to be male
- (surprisingly) people who report above average non-household contacts are slightly more likely to be college educated
- people who report above average non-household contacts are more likely to be employed
- people who report above average non-household contacts are more likely to be white
- people who report above average non-household contacts are slightly less likely to be Hispanic
- (surprisingly) people who report above average non-household contacts are those with higher household size
- people who report below average non-household contacts are more likely to come from dense urbanized settings
```{r}
average_by_contacts <- filtered_df %>%
  group_by(num_cc_nonhh > mean(num_cc_nonhh, na.rm = TRUE)) %>%
  filter(!is.na(Contacts)) %>% 
  summarise(
    average_age = mean(age, na.rm = TRUE),
    average_male = mean(male, na.rm = TRUE),
    average_college = mean(r_college_grad, na.rm = TRUE),
    average_employed = mean(r_working_num, na.rm = TRUE),
    average_white = mean(white, na.rm = TRUE),
    average_hispanic = mean(r_hispanic_num, na.rm = TRUE),
    average_household_size = mean(resp_hhsize, na.rm = TRUE),
    average_urban = mean(COUNTY_RUCC_2013, na.rm = TRUE)
  )

average_by_contacts
```
What about masked contacts (grouped by greater than half == TRUE)

High level findings:
- All but age are different 

Detailed Findings:
- There's not much difference in age between groups
- People who report mask usage more often are more likely to be male
- People who report mask usage more often are more likely to be college educated
- People who report mask usage more often are more likely to be employed
- People who report mask usage more often are less likely to be white
- People who report mask usage more often are slighly less likely to be Hispanic
- People who report mask usage more often are more likely to come from a larger household
- People who report mask usage more often more often from densely urbanized counties
```{r}
average_by_masks <- filtered_df %>%
  group_by(Norm_Masks_Used > mean(Norm_Masks_Used, na.rm = TRUE)) %>% #anyone who wore a mask in greater than half of their interactions
  filter(!is.na(Norm_Masks_Used)) %>% 
  summarise(
    average_age = mean(age, na.rm = TRUE),
    average_male = mean(male, na.rm = TRUE),
    average_college = mean(r_college_grad, na.rm = TRUE),
    average_employed = mean(r_working_num, na.rm = TRUE),
    average_white = mean(white, na.rm = TRUE),
    average_hispanic = mean(r_hispanic_num, na.rm = TRUE),
    average_household_size = mean(resp_hhsize, na.rm = TRUE),
    average_urban = mean(COUNTY_RUCC_2013, na.rm = TRUE)
  )

average_by_masks
```

What about democrats who are within republican districts?
High level findings:
- Democrats in majority republican CD's resemble them in a few ways. They are older, whiter and less Hispanic, less likely to have a college education, less likely to be in highly urbanized counties. However, they differ in that they're less likely to be employed, have smaller household size. Compared to Republicans, Democrats are much less likely to live in an opposing party district (could just be a result of how sample was collected). 
```{r}
filtered_df_D <- national_df %>% 
  filter(!is.na(CD_PERCENT_REPUBLICAN) & political_party == "Democrat")

average_by_CD_D <- filtered_df_D %>%
  group_by(CD_PERCENT_REPUBLICAN > .5) %>%
  summarise(
    percent = n()/nrow(filtered_df_D),
    average_age = mean(age, w = weight_pooled, na.rm = TRUE),
    average_male = mean(male, w = weight_pooled, na.rm = TRUE),
    average_college = mean(r_college_grad, w = weight_pooled, na.rm = TRUE),
    average_employed = mean(r_working_num, w = weight_pooled, na.rm = TRUE),
    average_white = mean(white, w = weight_pooled, na.rm = TRUE),
    average_hispanic = mean(r_hispanic_num, w = weight_pooled, na.rm = TRUE),
    average_household_size = mean(resp_hhsize, w = weight_pooled, na.rm = TRUE),
    average_urban = mean(COUNTY_RUCC_2013, w = weight_pooled, na.rm = TRUE),
    average_extremism = mean(extreme_liberal, w = weight_pooled, na.rm = TRUE)
  )

average_by_CD_D
```


And, what about republicans in democratic districts?
High level findings:
- Republicans in highly democratic areas resemble democrats. They're younger, they're more likely to have a college education, they're less white, more Hispanic, and more likely to live in a dense urban area. However, they're also likely to have a higher household size, more likely to be employed, more likely to be male. Republicans in highly democratic CD's are also less likely to be extreme conservatives. There's also a good amount of republicans in both categories. 
```{r}
filtered_df_R <- national_df %>% 
  filter(!is.na(CD_PERCENT_DEMOCRAT) & political_party == "Republican")

average_by_CD_R <- filtered_df_R %>%
  group_by(CD_PERCENT_DEMOCRAT > .5) %>%
  summarise(
    count = n()/nrow(filtered_df_R),
    average_age = mean(age, w = weight_pooled, na.rm = TRUE),
    average_male = mean(male, w = weight_pooled, na.rm = TRUE),
    average_college = mean(r_college_grad, w = weight_pooled, na.rm = TRUE),
    average_employed = mean(r_working_num, w = weight_pooled, na.rm = TRUE),
    average_white = mean(white, w = weight_pooled, na.rm = TRUE),
    average_hispanic = mean(r_hispanic_num, w = weight_pooled, na.rm = TRUE),
    average_household_size = mean(resp_hhsize, w = weight_pooled, na.rm = TRUE),
    average_urban = mean(COUNTY_RUCC_2013, w = weight_pooled, na.rm = TRUE),
    average_extremism = mean(extreme_conservative, w = weight_pooled, na.rm = TRUE)
  )

average_by_CD_R
```
Lastly, I want to control for county mask mandates in order to understand whether, for example, republicans in democratic districts aren't just limited their interactions and wearing more masks because of masks. However, are there any democratic counties without mask mandates? Are there any republicans in democratic counties that don't have mask mandates?

High level findings:
- Democrats are more likely to live in strict mask mandate counties, but both republicans and democrats are both not very likely to live in places with no mask mandate at all.
```{r}
mask_mandates <- bics_zip_features %>%
  filter(political_party != "Prefer not to answer" & political_party != "Independent") %>% 
  filter(!is.na(county_mask_mandate)) %>% 
  group_by(political_party, county_mask_mandate) %>%
  summarise(
    ount = n()
  )

mask_mandates
```
The group I'm most worried about is Republicans in majority democrat congressional districts without a county mask mandate.

High Level Findings:
- Although there are few, there are some republicans living in democratic-led CD's and in counties with no mask mandate (242)
```{r}
mask_mandates_R <- R_df %>%
  filter(!is.na(county_mask_mandate)) %>% 
  filter(!is.na(CD_PERCENT_DEMOCRAT)) %>% 
  group_by(CD_PERCENT_DEMOCRAT > .5, county_mask_mandate) %>%
  summarise(
    ount = n()
  )

mask_mandates_R
```














```{r}
rm(average_by_party, average_by_concern, filtered_df, average_by_contacts, average_by_masks, average_by_vaccine, filtered_df_D, filtered_df_R, average_by_CD_R, average_by_CD_D,mask_mandates,mask_mandates_R)
```

Next, I want to produce some simple bivariate results in order to show the magnitude of the effect
```{r}
bi_age <- lm(num_cc_nonhh ~ young, data = bics_zip_features, weights = weight_pooled)
bi_party <- lm(num_cc_nonhh ~ republican, data = RD_df, weights = weight_pooled)
bi_race <- lm(num_cc_nonhh ~ white, data = bics_zip_features, weights = weight_pooled)
bi_gender <- lm(num_cc_nonhh ~ male, data = bics_zip_features, weights = weight_pooled)
bi_concern <- lm(num_cc_nonhh ~ strong_concern_inverse, data = bics_zip_features, weights = weight_pooled)

summary(bi_party)

emm_race <- emmeans(bi_race, specs = pairwise ~ white)
emm_gender <- emmeans(bi_gender, specs = pairwise ~ male)
emm_age <- emmeans(bi_age, specs = pairwise ~ young)
emm_party <- emmeans(bi_party, specs = pairwise ~ republican)
emm_concern <- emmeans(bi_concern, specs = pairwise ~ strong_concern_inverse)

contr_race <- contrast(emm_race, "pairwise")
contr_gender <- contrast(emm_gender, "pairwise")
contr_age <- contrast(emm_age, "pairwise")
contr_party <- contrast(emm_party, "pairwise")
contr_concern <- contrast(emm_concern, "pairwise")

df_race_contrasts <- summary(contr_race) %>% as.data.frame()
df_gender_contrasts <- summary(contr_gender) %>% as.data.frame()
df_age_contrasts <- summary(contr_age) %>% as.data.frame()
df_party_contrasts <- summary(contr_party) %>% as.data.frame()
df_concern <- summary(contr_concern) %>% as.data.frame()

emm_race$variable <- "white"
emm_gender$variable <- "male"
emm_age$variable <- "65+"
emm_party$variable <- "republican"
emm_concern$variable <- "very concerned"

# Combine all emm_* data frames into one
combined_contrasts <- bind_rows(df_age_contrasts, df_party_contrasts, df_race_contrasts, df_gender_contrasts, df_concern)
combined_contrasts <- as.data.frame(combined_contrasts)

combined_contrasts$contrast <- factor(combined_contrasts$contrast, levels = unique(combined_contrasts$contrast))


pred_race <- ggplot(combined_contrasts, aes(x = contrast, y = -1 * estimate, fill = contrast)) +
  geom_bar(stat = "identity", position = position_dodge(), width = 0.7) +  # Draw bars
  geom_errorbar(aes(ymin = -1 * (estimate + SE), ymax = -1 * (estimate - SE)), 
                position = position_dodge(0.7), width = 0.25) +  # Add error bars
  labs(x = "", y = "", title = "differential non-household contacts by group") +
  scale_x_discrete(labels = c("Young - Old", "Republican - Democrat", "White - Non-White", "Male - Non-Male", "very concerned - not very concerned")) +
  scale_fill_manual(values = c("young0 - young1" = "gray", "republican0 - republican1" = "red", 
                               "white0 - white1" = "gray", "male0 - male1" = "gray", "very concerned - not very concerned" = "darkgreen")) +
  theme_minimal() +
  theme(legend.position = "none",
        plot.title = element_text(size = 18),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", size = 0.5),
        panel.grid.major.y = element_line(color = "grey90", size = 0.5),
        axis.title.x = element_text(size = 18),
        axis.text.x = element_text(size = 20),
        axis.text.y = element_text(size = 22),
        axis.title.y = element_text(size = 16),
        legend.title = element_text(size = 12),
        legend.text = element_text(size = 10))

ggsave("contrast_contacts.png", plot = pred_race, path = "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Contacts", width = 12, height = 8, units = "in")

pred_race

rm(emm_race, emm_gender, emm_age, emm_party, combined_contrasts, pred_race, df_race_contrasts, df_gender_contrasts, df_age_contrasts, df_party_contrasts, contr_age, contr_gender, contr_party, contr_race, bi_age, bi_gender, bi_party, bi_race)
```

```{r}
bi_age <- lm(Norm_Masks_Used ~ young, data = bics_zip_features, weights = weight_pooled)
bi_party <- lm(Norm_Masks_Used ~ republican, data = RD_df, weights = weight_pooled)
bi_race <- lm(Norm_Masks_Used ~ white, data = bics_zip_features, weights = weight_pooled)
bi_gender <- lm(Norm_Masks_Used ~ male, data = bics_zip_features, weights = weight_pooled)

emm_race <- emmeans(bi_race, specs = pairwise ~ white)
emm_gender <- emmeans(bi_gender, specs = pairwise ~ male)
emm_age <- emmeans(bi_age, specs = pairwise ~ young)
emm_party <- emmeans(bi_party, specs = pairwise ~ republican)

contr_race <- contrast(emm_race, "pairwise")
contr_gender <- contrast(emm_gender, "pairwise")
contr_age <- contrast(emm_age, "pairwise")
contr_party <- contrast(emm_party, "pairwise")

# Convert emmeans summaries to data frames
df_race_contrasts <- summary(contr_race) %>% as.data.frame()
df_gender_contrasts <- summary(contr_gender) %>% as.data.frame()
df_age_contrasts <- summary(contr_age) %>% as.data.frame()
df_party_contrasts <- summary(contr_party) %>% as.data.frame()

# Correct: Assign 'variable' column directly to emm_* data frames
emm_race$variable <- "white"
emm_gender$variable <- "male"
emm_age$variable <- "65+"
emm_party$variable <- "republican"

# Combine all emm_* data frames into one
combined_contrasts <- bind_rows(df_age_contrasts, df_party_contrasts, df_race_contrasts, df_gender_contrasts)
combined_contrasts <- as.data.frame(combined_contrasts)

combined_contrasts$contrast <- factor(combined_contrasts$contrast, levels = unique(combined_contrasts$contrast))


pred_race <- ggplot(combined_contrasts, aes(x = contrast, y = -1 * estimate, fill = contrast)) +
  geom_bar(stat = "identity", position = position_dodge(), width = 0.7) +  # Draw bars
  geom_errorbar(aes(ymin = -1 * (estimate + SE), ymax = -1 * (estimate - SE)), 
                position = position_dodge(0.7), width = 0.25) +  # Add error bars
  labs(x = "", y = "", title = "differential mask-usage by group") +
  scale_x_discrete(labels = c("Young - Old", "Republican - Democrat", "White - Non-White", "Male - Non-Male")) +
  scale_fill_manual(values = c("young0 - young1" = "gray", "republican0 - republican1" = "red", 
                               "white0 - white1" = "gray", "male0 - male1" = "gray")) +
  theme_minimal() +
  theme(legend.position = "none",
        plot.title = element_text(size = 18),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", size = 0.5),
        panel.grid.major.y = element_line(color = "grey90", size = 0.5),
        axis.title.x = element_text(size = 18),
        axis.text.x = element_text(size = 20),
        axis.text.y = element_text(size = 22),
        axis.title.y = element_text(size = 16),
        legend.title = element_text(size = 12),
        legend.text = element_text(size = 10))

ggsave("contrast_masks.png", plot = pred_race, path = "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Masks", width = 12, height = 8, units = "in")

pred_race

rm(emm_race, emm_gender, emm_age, emm_party, combined_contrasts, pred_race, df_race_contrasts, df_gender_contrasts, df_age_contrasts, df_party_contrasts, contr_age, contr_gender, contr_party, contr_race, bi_age, bi_gender, bi_party, bi_race)
```
Let's find out how prevalence rates impact contacts and mask usage. Previous week rates are also significant predictors, but not as strong as current week leading up to data collection date
```{r}
bi_nonhh <- lm(num_cc_nonhh ~ log_current_week_inc_rate, data = bics_zip_features, weights = weight_pooled)
bi_mask <- lm(Norm_Masks_Used ~ log_current_week_inc_rate, data = bics_zip_features, weights = weight_pooled)
print(summary(bi_nonhh))
print(summary(bi_mask))
bi_nonhh_lag <- lm(num_cc_nonhh ~ log_prev_week_inc_rate, data = bics_zip_features, weights = weight_pooled)
bi_mask_lag <- lm(Norm_Masks_Used ~ log_prev_week_inc_rate, data = bics_zip_features, weights = weight_pooled)
print(summary(bi_nonhh_lag))
print(summary(bi_mask_lag))

rm(bi_nonhh,bi_mask,bi_nonhh_lag,bi_mask_lag)
```
I find that the association between incidence and non-household continues when I add in a control for wave. 
```{r}
bi_nonhh <- lm(num_cc_nonhh ~ log_current_week_inc_rate + wave, data = bics_zip_features, weights = weight_pooled)
bi_mask <- lm(Norm_Masks_Used ~ log_current_week_inc_rate + wave, data = bics_zip_features, weights = weight_pooled)
print(summary(bi_nonhh))
print(summary(bi_mask))
bi_nonhh_lag <- lm(num_cc_nonhh ~ log_prev_week_inc_rate + wave, data = bics_zip_features, weights = weight_pooled)
bi_mask_lag <- lm(Norm_Masks_Used ~ log_prev_week_inc_rate + wave, data = bics_zip_features, weights = weight_pooled)
print(summary(bi_nonhh_lag))
print(summary(bi_mask_lag))

rm(bi_nonhh,bi_mask,bi_nonhh_lag,bi_mask_lag)
```
Let's see at which wave the association breaks down. 
```{r}
contacts_R_v_D_w2 <- lm(num_cc_nonhh ~  log_current_week_inc_rate, data = subset(bics_zip_features, wave == 2, weights = weight_pooled))

#model for Wave 3 (September 12-20, '20)
contacts_R_v_D_w3 <- lm(num_cc_nonhh ~ log_current_week_inc_rate, data = subset(bics_zip_features, wave == 3, weights = weight_pooled))

# Model for Wave 4 (December 1-7, '20')
contacts_R_v_D_w4 <- lm(num_cc_nonhh ~  log_current_week_inc_rate, data = subset(bics_zip_features, wave == 4, weights = weight_pooled))

#model for Wave 5 (February 9-14, '21')
contacts_R_v_D_w5 <- lm(num_cc_nonhh ~ log_current_week_inc_rate, data = subset(bics_zip_features, wave == 5, weights = weight_pooled))

# Model for Wave 6 (May 14-25, '21')
contacts_R_v_D_w6 <- lm(num_cc_nonhh ~  log_current_week_inc_rate, data = subset(bics_zip_features, wave == 6, weights = weight_pooled))

print(summary(contacts_R_v_D_w2))

print(summary(contacts_R_v_D_w3))

print(summary(contacts_R_v_D_w4))

print(summary(contacts_R_v_D_w5))

print(summary(contacts_R_v_D_w6))

rm(contacts_R_v_D_w2,contacts_R_v_D_w3,contacts_R_v_D_w4,contacts_R_v_D_w5,contacts_R_v_D_w6)
```


Is the incidence rate a stronger predictor of democrat behavior than it is republican?
Interestingly, incidence is only predictive of dem contact behavior but not republican
```{r}
bi_nonhh_dem <- lm(num_cc_nonhh ~ log_current_week_inc_rate, data = D_df, weights = weight_pooled)
bi_nonhh_rep <- lm(num_cc_nonhh ~ log_current_week_inc_rate, data = R_df, weights = weight_pooled)

print(summary(bi_nonhh_dem))
print(summary(bi_nonhh_rep))

rm(bi_nonhh_dem,bi_nonhh_rep)
```
And mask usage?
Democrats are more reactive to incidence rates than Republicans are with a coefficient twice as large. In other words, every .02 increase in the incidence rate is associated with a 1% increase in the percentage of contacts that are carried out with a mask. 
```{r}
bi_mask_dem <- lm(Norm_Masks_Used ~ log_current_week_inc_rate, data = D_df, weights = weight_pooled)
bi_mask_rep <- lm(Norm_Masks_Used ~ log_current_week_inc_rate, data = R_df, weights = weight_pooled)

print(summary(bi_mask_dem))
print(summary(bi_mask_rep))

rm(bi_mask_dem,bi_mask_rep)
```


Lastly, let's do vaccination
```{r}
bi_age <- lm(Vaccinated ~ young, data = bics_zip_features, weights = weight_pooled)
bi_party <- lm(Vaccinated ~ republican, data = RD_df, weights = weight_pooled)
bi_race <- lm(Vaccinated ~ white, data = bics_zip_features, weights = weight_pooled)
bi_gender <- lm(Vaccinated ~ male, data = bics_zip_features, weights = weight_pooled)

summary(bi_party)

emm_race <- emmeans(bi_race, specs = pairwise ~ white)
emm_gender <- emmeans(bi_gender, specs = pairwise ~ male)
emm_age <- emmeans(bi_age, specs = pairwise ~ young)
emm_party <- emmeans(bi_party, specs = pairwise ~ republican)

contr_race <- contrast(emm_race, "pairwise")
contr_gender <- contrast(emm_gender, "pairwise")
contr_age <- contrast(emm_age, "pairwise")
contr_party <- contrast(emm_party, "pairwise")

# Convert emmeans summaries to data frames
df_race_contrasts <- summary(contr_race) %>% as.data.frame()
df_gender_contrasts <- summary(contr_gender) %>% as.data.frame()
df_age_contrasts <- summary(contr_age) %>% as.data.frame()
df_party_contrasts <- summary(contr_party) %>% as.data.frame()

# Correct: Assign 'variable' column directly to emm_* data frames
emm_race$variable <- "white"
emm_gender$variable <- "male"
emm_age$variable <- "65+"
emm_party$variable <- "republican"

# Combine all emm_* data frames into one
combined_contrasts <- bind_rows(df_age_contrasts, df_party_contrasts, df_race_contrasts, df_gender_contrasts)
combined_contrasts <- as.data.frame(combined_contrasts)

combined_contrasts$contrast <- factor(combined_contrasts$contrast, levels = unique(combined_contrasts$contrast))


pred_race <- ggplot(combined_contrasts, aes(x = contrast, y = -1 * estimate, fill = contrast)) +
  geom_bar(stat = "identity", position = position_dodge(), width = 0.7) + 
  geom_errorbar(aes(ymin = -1 * (estimate + SE), ymax = -1 * (estimate - SE)), 
                position = position_dodge(0.7), width = 0.25) + 
  labs(x = "", y = "", title = "differential vaccination rates by group") +
  scale_x_discrete(labels = c("Young-Old", "Republican-Democrat", "White-Non-White", "Male-Non-Male")) +
  scale_fill_manual(values = c("young0 - young1" = "gray", "republican0 - republican1" = "red", 
                               "white0 - white1" = "gray", "male0 - male1" = "gray")) +
  theme_minimal() +
  theme(legend.position = "none",
        plot.title = element_text(size = 18),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", size = 0.5),
        panel.grid.major.y = element_line(color = "grey90", size = 0.5),
        axis.title.x = element_text(size = 18),
        axis.text.x = element_text(size = 20),
        axis.text.y = element_text(size = 20),
        axis.title.y = element_text(size = 16),
        legend.title = element_text(size = 12),
        legend.text = element_text(size = 10)) 

ggsave("contrast_vaccines.png", plot = pred_race, path = "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Vaccines", width = 12, height = 8, units = "in")

pred_race

rm(emm_race, emm_gender, emm_age, emm_party, combined_contrasts, pred_race, df_race_contrasts, df_gender_contrasts, df_age_contrasts, df_party_contrasts, contr_age, contr_gender, contr_party, contr_race, bi_age, bi_gender, bi_party, bi_race)
```

```{r}
contact_reduction_R_v_D_w2 <- lm(num_cc_nonhh ~  log_rate_prev_6_14_20 + political_party, data = subset(bics_zip_features, wave == 2))

# Model for Wave 3 (September 12-20, '20)
contact_reduction_R_v_D_w3 <- lm(num_cc_nonhh ~ log_rate_prev_9_9_20 + political_party, data = subset(bics_zip_features, wave == 3))

# Model for Wave 4 (December 1-7, '20)
contact_reduction_R_v_D_w4 <- lm(num_cc_nonhh ~  log_rate_prev_12_01_20 + political_party, data = subset(bics_zip_features, wave == 4))

# Model for Wave 5 (February 9-14, '21)
contact_reduction_R_v_D_w5 <- lm(num_cc_nonhh ~ log_rate_prev_2_4_21 + political_party, data = subset(bics_zip_features, wave == 5))

# Model for Wave 6 (May 14-25, '21)
contact_reduction_R_v_D_w6 <- lm(num_cc_nonhh ~  log_rate_prev_5_12_21 + political_party, data = subset(bics_zip_features, wave == 6))

emm_w2 <- emmeans(contact_reduction_R_v_D_w2, specs = "political_party")
emm_w3 <- emmeans(contact_reduction_R_v_D_w3, specs = "political_party")
emm_w4 <- emmeans(contact_reduction_R_v_D_w4, specs = "political_party")
emm_w5 <- emmeans(contact_reduction_R_v_D_w5, specs = "political_party")
emm_w6 <- emmeans(contact_reduction_R_v_D_w6, specs = "political_party")

all_emms <- rbind(
  transform(summary(emm_w2), wave = "June 18-24, '20"),
  transform(summary(emm_w3), wave = "September 12-20, '20"),
  transform(summary(emm_w4), wave = "December 1-7, '20"),
  transform(summary(emm_w5), wave = "February 9-14, '21"),
  transform(summary(emm_w6), wave = "May 14-25, '21")
)

rm(emm_w2, emm_w3, emm_w4, emm_w5, emm_w6)
rm(contact_reduction_R_v_D_w2,contact_reduction_R_v_D_w3,contact_reduction_R_v_D_w4,contact_reduction_R_v_D_w5,contact_reduction_R_v_D_w6)

all_emms <- all_emms %>% 
  filter(political_party != "Prefer not to answer")

all_emms$wave <- factor(all_emms$wave, levels = cases_df$wave)

contact_reduction_rvd_waves <- ggplot(all_emms, aes(x = wave, y = emmean, color = political_party, shape = political_party)) +
  geom_point(position = position_dodge(width = 0.35), size = 3) +
  labs(y = "Predicted Reported Contacts", x = "", color = "Political Party") +
  geom_errorbar(aes(ymin = emmean - SE, ymax = emmean + SE), width = 0.2,position = position_dodge(width = 0.35)) +
  scale_color_manual(name = "emmean", 
                     values = c("blue", "red", "darkgreen")) +
  theme_minimal()

contact_reduction_rvd_waves

rm(all_emms,contact_reduction_rvd_waves)
```
Lastly, I want to get a sense of the most and least polarized counties, see how polarization changes over time, and get a sense of what drives polarization
let's start with polarization over time

```{r}
library(dplyr)
library(purrr)

# function
summarize_data <- function(df, wave_num, group) {
  df %>%
    filter(wave == wave_num) %>%
    summarise(
      wave = wave_num,
      num_non_houshold_cc = mean(num_cc_nonhh, na.rm = TRUE),
      num_cc = mean(num_cc, w = weight_pooled, na.rm = TRUE),
      percent_mask = mean(Norm_Masks_Used, na.rm = TRUE),
      vaccinated = mean(Vaccinated, na.rm = TRUE),
      concern =mean(binary_concern_strong, na.rm = TRUE)
    ) %>%
    mutate(group = group)
}

datasets <- list(D = D_df, R = R_df)
waves <- c(2, 3, 4, 5, 6)

results <- list()

for (name in names(datasets)) {
  for (wave_num in waves) {
    result_name <- paste("average_for", name, "w", wave_num, sep = "_")
    results[[result_name]] <- summarize_data(datasets[[name]], wave_num, name)
  }
}

combined_results <- bind_rows(results)

wide_results <- combined_results %>%
  pivot_wider(names_from = group, values_from = c(num_non_houshold_cc, num_cc, percent_mask, vaccinated, concern)) %>%
  rename(
    D_num_non_houshold_cc = num_non_houshold_cc_D,
    R_num_non_houshold_cc = num_non_houshold_cc_R,
    D_num_cc = num_cc_D,
    R_num_cc = num_cc_R,
    D_percent_mask = percent_mask_D,
    R_percent_mask = percent_mask_R,
    D_vaccinated = vaccinated_D,
    R_vaccinated = vaccinated_R,
    D_concern = concern_D,
    R_concern = concern_R
  )

differences_df <- data.frame(
  wave = wide_results$wave,
  diff_num_non_houshold_cc = wide_results$R_num_non_houshold_cc - wide_results$D_num_non_houshold_cc,
  diff_num_cc = wide_results$R_num_cc - wide_results$D_num_cc,
  diff_percent_mask = wide_results$R_percent_mask - wide_results$D_percent_mask,
  diff_vaccinated = wide_results$R_vaccinated - wide_results$D_vaccinated,
  diff_concern = wide_results$R_concern - wide_results$D_concern,
  abs_diff_num_non_houshold_cc = abs(wide_results$R_num_non_houshold_cc - wide_results$D_num_non_houshold_cc),
  abs_diff_num_cc = abs(wide_results$R_num_cc - wide_results$D_num_cc),
  abs_diff_percent_mask = abs(wide_results$R_percent_mask - wide_results$D_percent_mask),
  abs_diff_vaccinated = abs(wide_results$R_vaccinated - wide_results$D_vaccinated),
  abs_diff_concern = abs(wide_results$R_concern - wide_results$D_concern)
)

print(differences_df)

differences_df$weekly_cases <-cases_df$weekly_cases
differences_df$dates <-cases_df$formal_dates
```

```{r}
ggplot(differences_df, aes(x = wave, y = abs_diff_num_non_houshold_cc)) +
  geom_line() +
  geom_point() +
  labs(
    title = "Difference in Non-household CC by Wave",
    x = "Wave",
    y = "Difference in Non-household CC"
  ) +
  theme_minimal() +
  theme(
    plot.title = element_text(hjust = 0.5)
  )

ggplot(differences_df, aes(x = wave, y = abs_diff_percent_mask)) +
  geom_line() +
  geom_point() +
  labs(
    title = "Difference in Percent Masks Used by Wave",
    x = "Wave",
    y = "Difference in Non-household CC"
  ) +
  theme_minimal() +
  theme(
    plot.title = element_text(hjust = 0.5)
  )

ggplot(differences_df, aes(x = wave, y = abs_diff_concern)) +
  geom_line() +
  geom_point() +
  labs(
    title = "Difference in Concern Used by Wave",
    x = "Wave",
    y = "Difference in Non-household CC"
  ) +
  theme_minimal() +
  theme(
    plot.title = element_text(hjust = 0.5)
  )
```

here's another set of plots to overlay cases. Only meant to show how the two are related, and communicate nothing else other than visualizing their movement.
```{r}
# non-household contact
ggplot(differences_df, aes(x = wave)) +
  geom_bar(aes(y = weekly_cases), stat = "identity", fill = "skyblue", alpha = 0.6) +
  geom_line(aes(y = abs_diff_num_non_houshold_cc * 1000), color = "purple", size = 1) + 
  geom_point(aes(y = abs_diff_num_non_houshold_cc * 1000), color = "purple", size = 3) +
  scale_y_continuous(
    name = "Weekly Cases",
    sec.axis = sec_axis(~./10000, name = "Difference in Non-household CC")
  ) +
  labs(
    title = "Difference in Non-household CC and Weekly Cases by Wave",
    x = "Wave"
  ) +
  theme_minimal() +
  theme(
    plot.title = element_text(hjust = 0.5)
  )

#masks
ggplot(differences_df, aes(x = wave)) +
  geom_bar(aes(y = weekly_cases), stat = "identity", fill = "skyblue", alpha = 0.6) +
  geom_line(aes(y = abs_diff_percent_mask * 100), color = "purple", size = 1) + 
  geom_point(aes(y = abs_diff_percent_mask * 100), color = "purple", size = 3) + 
  scale_y_continuous(
    name = "Weekly Cases",
    sec.axis = sec_axis(~./10000, name = "Difference in Mask Usage") 
  ) +
  labs(
    title = "Difference in Mask Usage and Weekly Cases by Wave",
    x = "Wave"
  ) +
  theme_minimal() +
  theme(
    plot.title = element_text(hjust = 0.5)
  )

# concern
ggplot(differences_df, aes(x = wave)) +
  geom_bar(aes(y = weekly_cases), stat = "identity", fill = "skyblue", alpha = 0.6) +
  geom_line(aes(y = abs_diff_concern * 10000), color = "purple", size = 1) +
  geom_point(aes(y = abs_diff_concern * 10000), color = "purple", size = 3) + 
  scale_y_continuous(
    name = "Weekly Cases",
    sec.axis = sec_axis(~./10000, name = "Difference in Concern About Spread")
  ) +
  labs(
    title = "Difference in Concern about Spread and Weekly Cases by Wave",
    x = "Wave"
  ) +
  theme_minimal() +
  theme(
    plot.title = element_text(hjust = 0.5) 
  )
```

Let's plot the incidence rate alongside waves for descriptive purposes
```{r}
cases_plot <- ggplot(cases_df, aes(x = formal_dates, y = weekly_cases)) +
  geom_col(fill = "darkgray", width = 25) + 
  labs(x = "", y = "",title = "Weekly Average Cases") +
  theme_minimal() +
  scale_x_date(breaks = cases_df$formal_dates, labels = scales::date_format("%b, '%y")) + 
  theme(legend.position = "none",
        plot.title = element_text(size = 18),
        panel.grid.minor = element_blank(),
        panel.grid.major = element_blank(),
        axis.title.x = element_text(size = 18),
        axis.text.x = element_text(size = 25),
        axis.text.y = element_text(size = 20),
        axis.title.y = element_text(size = 16),
        legend.title = element_text(size = 12),
        legend.text = element_text(size = 10))

print(cases_plot)

ggsave("cases.png", plot = cases_plot, path = "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/", width = 12, height = 8, units = "in")
```
Can we make any assumptions about homophily of contacts between republicans and democrats?
Paper by Brown and Enos suggests that Democrats are more homophilous than Republicans
https://www.nature.com/articles/s41562-021-01066-z
They find that, on average,
