==> code/05-rep_v_dem_waves.Rmd <==
---
title: "04-rep_v_dem_waves"
author: "Christopher Soria"
date: "2024-02-08"
output: html_document
---
```{r}
library(emmeans)
emm_options(rg.limit = 20000)
library(tidyverse)
```
contact reduction
```{r}
contact_reduction_R_v_D_w2 <- lm(contact_reduction ~  log_rate_prev_6_14_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 2))

# Model for Wave 3 (September 12-20, '20)
contact_reduction_R_v_D_w3 <- lm(contact_reduction ~ log_rate_prev_9_9_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 3))

# Model for Wave 4 (December 1-7, '20)
contact_reduction_R_v_D_w4 <- lm(contact_reduction ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 4))

# Model for Wave 5 (February 9-14, '21)
contact_reduction_R_v_D_w5 <- lm(contact_reduction ~ log_rate_prev_2_4_21 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 5))

# Model for Wave 6 (May 14-25, '21)
contact_reduction_R_v_D_w6 <- lm(contact_reduction ~  log_rate_prev_5_12_21 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 6))

emm_w2 <- emmeans(contact_reduction_R_v_D_w2, specs = "political_party")
emm_w3 <- emmeans(contact_reduction_R_v_D_w3, specs = "political_party")
emm_w4 <- emmeans(contact_reduction_R_v_D_w4, specs = "political_party")
emm_w5 <- emmeans(contact_reduction_R_v_D_w5, specs = "political_party")
emm_w6 <- emmeans(contact_reduction_R_v_D_w6, specs = "political_party")

all_emms <- rbind(
  transform(summary(emm_w2), wave = "June 18-24, '20"),
  transform(summary(emm_w3), wave = "September 12-20, '20"),
  transform(summary(emm_w4), wave = "December 1-7, '20"),
  transform(summary(emm_w5), wave = "February 9-14, '21"),
  transform(summary(emm_w6), wave = "May 14-25, '21")
)

rm(contact_reduction_R_v_D_w2, contact_reduction_R_v_D_w3, contact_reduction_R_v_D_w4, contact_reduction_R_v_D_w5, contact_reduction_R_v_D_w6)
rm(emm_w2, emm_w3, emm_w4, emm_w5, emm_w6)

all_emms <- all_emms %>% 
  filter(political_party != "Prefer not to answer")

all_emms$wave <- factor(all_emms$wave, levels = cases_df$wave)

contact_reduction_rvd_waves <- ggplot(all_emms, aes(x = wave, y = emmean, color = political_party, shape = political_party)) +
  geom_point(position = position_dodge(width = 0.35), size = 3) +
  labs(y = "Predicted Reported Contacts", x = "", color = "Political Party") +
  geom_errorbar(aes(ymin = emmean - SE, ymax = emmean + SE), width = 0.2,position = position_dodge(width = 0.35)) +
  scale_color_manual(name = "emmean", 
                     values = c("blue", "red", "darkgreen")) +
  theme_minimal()

ggsave("rvd_contact_reduction_waves.png", plot = contact_reduction_rvd_waves, path = "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Contact_Reduction", width = 12, height = 8, units = "in")

contact_reduction_rvd_waves

rm(all_emms,contact_reduction_rvd_waves)
```
stated concern
```{r}
# Model for Wave 2 (June 20-30, '20)
concern_R_v_D_w2 <- lm(binary_concern_strong ~  log_rate_prev_6_14_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 2))

# Model for Wave 3 (September 12-20, '20)
concern_R_v_D_w3 <- lm(binary_concern_strong ~ log_rate_prev_9_9_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 3))

# Model for Wave 4 (December 1-7, '20)
concern_R_v_D_w4 <- lm(binary_concern_strong ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 4))

# Model for Wave 5 (February 9-14, '21)
concern_R_v_D_w5 <- lm(binary_concern_strong ~ log_rate_prev_2_4_21 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 5))

# Model for Wave 6 (May 14-25, '21)
concern_R_v_D_w6 <- lm(binary_concern_strong ~  log_rate_prev_5_12_21 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 6))

emm_w2 <- emmeans(concern_R_v_D_w2, specs = "political_party")
emm_w3 <- emmeans(concern_R_v_D_w3, specs = "political_party")
emm_w4 <- emmeans(concern_R_v_D_w4, specs = "political_party")
emm_w5 <- emmeans(concern_R_v_D_w5, specs = "political_party")
emm_w6 <- emmeans(concern_R_v_D_w6, specs = "political_party")

rm(concern_R_v_D_w2, concern_R_v_D_w3, concern_R_v_D_w4, concern_R_v_D_w5, concern_R_v_D_w6)

# Combine EMMs from all models and label wave collection dates
all_emms <- rbind(
  transform(summary(emm_w2), wave = "June 18-24, '20"),
  transform(summary(emm_w3), wave = "September 12-20, '20"),
  transform(summary(emm_w4), wave = "December 1-7, '20"),
  transform(summary(emm_w5), wave = "February 9-14, '21"),
  transform(summary(emm_w6), wave = "May 14-25, '21")
)

rm(emm_w2, emm_w3, emm_w4, emm_w5, emm_w6)

all_emms <- all_emms %>% 
  filter(political_party != "Prefer not to answer")

all_emms$wave <- factor(all_emms$wave, levels = cases_df$wave)

concern_rvd_waves <- ggplot(all_emms, aes(x = wave, y = emmean, color = political_party, shape = political_party)) +
  geom_point(position = position_dodge(width = 0.35), size = 4) +
  scale_color_manual(name = "Political Party", values = c("blue", "red", "darkgreen")) +
  labs(y = "Probability of Concern", x = "", color = "Political Party", title = "Predicted Concern About the Spread of the Virus Over Time") +  # Make sure to add your title here
  geom_errorbar(aes(ymin = emmean - SE, ymax = emmean + SE), width = 0.2, position = position_dodge(width = 0.35)) +
  scale_x_discrete(labels = c("Jun 18-24, '20", "Sept 12-20, '20", "Dec 1-7, '20", "Feb 9-14, '21", "May 14-25, '21")) +
  theme_minimal() +
  theme(plot.title = element_text(size = 26),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", size = 0.5),
        panel.grid.major.y = element_line(color = "grey90", size = 0.5),
        axis.title.x = element_text(size = 18),
        axis.text.x = element_text(size = 15),
        axis.title.y = element_text(size = 16),
        legend.title = element_text(size = 12),
        legend.text = element_text(size = 10)) +
  guides(shape = "none")

ggsave("rvd_concern_waves.png", plot = concern_rvd_waves, path = "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Concern", width = 12, height = 8, units = "in")

concern_rvd_waves

rm(all_emms,concern_rvd_waves)
```

contacts
```{r}
# Model for Wave 2 (June 18-24, '20)
contacts_R_v_D_w2 <- lm(num_cc_nonhh ~  current_week_inc_rate + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party + r_working, data = subset(bics_zip_features, wave == 2, weights = weight_pooled))

#model for Wave 3 (September 12-20, '20)
contacts_R_v_D_w3 <- lm(num_cc_nonhh ~ current_week_inc_rate + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party + r_working, data = subset(bics_zip_features, wave == 3))

# Model for Wave 4 (December 1-7, '20')
contacts_R_v_D_w4 <- lm(num_cc_nonhh ~  current_week_inc_rate + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party + r_working, data = subset(bics_zip_features, wave == 4))

#model for Wave 5 (February 9-14, '21')
contacts_R_v_D_w5 <- lm(num_cc_nonhh ~ current_week_inc_rate + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party + r_working, data = subset(bics_zip_features, wave == 5))

# Model for Wave 6 (May 14-25, '21')
contacts_R_v_D_w6 <- lm(num_cc_nonhh ~  current_week_inc_rate + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party + r_working, data = subset(bics_zip_features, wave == 6))

emm_w2 <- emmeans(contacts_R_v_D_w2, specs = "political_party")
emm_w3 <- emmeans(contacts_R_v_D_w3, specs = "political_party")
emm_w4 <- emmeans(contacts_R_v_D_w4, specs = "political_party")
emm_w5 <- emmeans(contacts_R_v_D_w5, specs = "political_party")
emm_w6 <- emmeans(contacts_R_v_D_w6, specs = "political_party")

# Combine EMMs from all models and label wave collection dates
all_emms <- rbind(
  transform(summary(emm_w2), wave = "June 18-24, '20"),
  transform(summary(emm_w3), wave = "September 12-20, '20"),
  transform(summary(emm_w4), wave = "December 1-7, '20"),
  transform(summary(emm_w5), wave = "February 9-14, '21"),
  transform(summary(emm_w6), wave = "May 14-25, '21")
)

all_emms <- all_emms %>% 
  filter(political_party != "Prefer not to answer")

all_emms$wave <- factor(all_emms$wave, levels = cases_df$wave)

rm(emm_w2, emm_w3, emm_w4, emm_w5, emm_w6)
rm(contacts_R_v_D_w2, contacts_R_v_D_w3, contacts_R_v_D_w4, contacts_R_v_D_w5, contacts_R_v_D_w6)

contacts_rvd_waves <- ggplot(all_emms, aes(x = wave, y = emmean, color = political_party, shape = political_party)) +
  geom_point(position = position_dodge(width = 0.35), size = 3) +
  labs(y = "Predicted Reported Contacts", x = "", color = "Political Party", title = "Predicted Non-Household Contacts Over Time") +
  geom_errorbar(aes(ymin = emmean - SE, ymax = emmean + SE), width = 0.2,position = position_dodge(width = 0.35)) +
  scale_color_manual(name = "Political Affiliation", 
                     values = c("blue", "red", "darkgreen")) +
  theme_minimal() +
  guides(shape = "none")

ggsave("rvd_contacts_waves.png", plot = contacts_rvd_waves, path = "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Contacts", width = 12, height = 8, units = "in")

contacts_rvd_waves
rm(all_emms,contacts_rvd_waves)
```
I wonder if this looks different in the national sample? The patterns are exactly the same but with larger error bars. I'll keep the pooled sample for the smaller error bar. 
```{r}
# Model for Wave 2 (June 18-24, '20)
contacts_R_v_D_w2 <- lm(num_cc_nonhh ~  log_rate_prev_6_14_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party + r_working, data = subset(national_df, wave == 2, weights = weight_pooled))

#model for Wave 3 (September 12-20, '20)
contacts_R_v_D_w3 <- lm(num_cc_nonhh ~ log_rate_prev_9_9_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party + r_working, data = subset(national_df, wave == 3))

# Model for Wave 4 (December 1-7, '20')
contacts_R_v_D_w4 <- lm(num_cc_nonhh ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party + r_working, data = subset(national_df, wave == 4))

#model for Wave 5 (February 9-14, '21')
contacts_R_v_D_w5 <- lm(num_cc_nonhh ~ log_rate_prev_2_4_21 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party + r_working, data = subset(national_df, wave == 5))

# Model for Wave 6 (May 14-25, '21')
contacts_R_v_D_w6 <- lm(num_cc_nonhh ~  log_rate_prev_5_12_21 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party + r_working, data = subset(national_df, wave == 6))

emm_w2 <- emmeans(contacts_R_v_D_w2, specs = "political_party")
emm_w3 <- emmeans(contacts_R_v_D_w3, specs = "political_party")
emm_w4 <- emmeans(contacts_R_v_D_w4, specs = "political_party")
emm_w5 <- emmeans(contacts_R_v_D_w5, specs = "political_party")
emm_w6 <- emmeans(contacts_R_v_D_w6, specs = "political_party")

# Combine EMMs from all models and label wave collection dates
all_emms <- rbind(
  transform(summary(emm_w2), wave = "June 18-24, '20"),
  transform(summary(emm_w3), wave = "September 12-20, '20"),
  transform(summary(emm_w4), wave = "December 1-7, '20"),
  transform(summary(emm_w5), wave = "February 9-14, '21"),
  transform(summary(emm_w6), wave = "May 14-25, '21")
)

all_emms <- all_emms %>% 
  filter(political_party != "Prefer not to answer")

all_emms$wave <- factor(all_emms$wave, levels = cases_df$wave)

rm(emm_w2, emm_w3, emm_w4, emm_w5, emm_w6)
rm(contacts_R_v_D_w2, contacts_R_v_D_w3, contacts_R_v_D_w4, contacts_R_v_D_w5, contacts_R_v_D_w6)

contacts_rvd_waves <- ggplot(all_emms, aes(x = wave, y = emmean, color = political_party, shape = political_party)) +
  geom_point(position = position_dodge(width = 0.35), size = 4) +
  scale_color_manual(name = "Political Party", values = c("blue", "red", "darkgreen")) +
  labs(y = "Predicted Reported Contacts", x = "", color = "Political Party", title = "Predicted Number of Non-Household Contacts Over Time") +  # Make sure to add your title here
  geom_errorbar(aes(ymin = emmean - SE, ymax = emmean + SE), width = 0.2, position = position_dodge(width = 0.35)) +
  scale_x_discrete(labels = c("Jun 18-24, '20", "Sept 12-20, '20", "Dec 1-7, '20", "Feb 9-14, '21", "May 14-25, '21")) +
  theme_minimal() +
  theme(plot.title = element_text(size = 26),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", size = 0.5),
        panel.grid.major.y = element_line(color = "grey90", size = 0.5),
        axis.title.x = element_text(size = 18),
        axis.text.x = element_text(size = 15),
        axis.title.y = element_text(size = 16),
        legend.title = element_text(size = 12),
        legend.text = element_text(size = 10)) +
  guides(shape = "none")

ggsave("rvd_contacts_waves_national.png", plot = contacts_rvd_waves, path = "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Contacts", width = 12, height = 8, units = "in")

contacts_rvd_waves
rm(all_emms,contacts_rvd_waves)
```


```{r}
cases_plot <- ggplot(cases_df, aes(x = wave, y = weekly_cases)) +
  geom_line(group = 1) +  # Add a line
  geom_point() +  # Add points
  labs(x = "Date", y = "Weekly Cases", title = "Weekly Average Deaths Over Time") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1))

ggsave("cases_waves.png", plot = cases_plot, path = "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/", width = 12, height = 8, units = "in")

cases_plot

rm(cases_plot)
```


masked contacts
```{r}
# Model for Wave 2 (June 20-30, '20)
masks_R_v_D_w2 <- lm(Norm_Masks_Used ~  log_rate_prev_6_14_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 2))

# Model for Wave 3 (September 12-20, '20)
masks_R_v_D_w3 <- lm(Norm_Masks_Used ~ log_rate_prev_9_9_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 3))

# Model for Wave 4 (December 1-7, '20)
masks_R_v_D_w4 <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 4))

# Model for Wave 5 (February 9-14, '21)
masks_R_v_D_w5 <- lm(Norm_Masks_Used ~ log_rate_prev_2_4_21 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 5))

# Model for Wave 6 (May 14-25, '21)
masks_R_v_D_w6 <- lm(Norm_Masks_Used ~  log_rate_prev_5_12_21 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + resp_hhsize + r_race + political_party, data = subset(bics_zip_features, wave == 6))

emm_w2 <- emmeans(masks_R_v_D_w2, specs = "political_party")
emm_w3 <- emmeans(masks_R_v_D_w3, specs = "political_party")
emm_w4 <- emmeans(masks_R_v_D_w4, specs = "political_party")
emm_w5 <- emmeans(masks_R_v_D_w5, specs = "political_party")
emm_w6 <- emmeans(masks_R_v_D_w6, specs = "political_party")

rm(masks_R_v_D_w2, masks_R_v_D_w3, masks_R_v_D_w4,masks_R_v_D_w5, masks_R_v_D_w6)

# Combine EMMs from all models and label wave collection dates
all_emms <- rbind(
  transform(summary(emm_w2), wave = "June 18-24, '20"),
  transform(summary(emm_w3), wave = "September 12-20, '20"),
  transform(summary(emm_w4), wave = "December 1-7, '20"),
  transform(summary(emm_w5), wave = "February 9-14, '21"),
  transform(summary(emm_w6), wave = "May 14-25, '21")
)

all_emms <- all_emms %>% 
  filter(political_party != "Prefer not to answer")

all_emms$wave <- factor(all_emms$wave, levels = cases_df$wave)

rm(emm_w2, emm_w3, emm_w4, emm_w5, emm_w6)

masks_rvd_waves <- ggplot(all_emms, aes(x = wave, y = emmean, color = political_party, shape = political_party)) +
  geom_point(position = position_dodge(width = 0.35), size = 4) +
  scale_color_manual(name = "Political Party", values = c("blue", "red", "darkgreen")) +
  labs(y = "Proportion of Contacts with a Mask", x = "", color = "Political Party", title = "Predicted Proportion of Contacts Carried Out With a Mask Over Time") +  # Make sure to add your title here
  geom_errorbar(aes(ymin = emmean - SE, ymax = emmean + SE), width = 0.2, position = position_dodge(width = 0.35)) +
  scale_x_discrete(labels = c("Jun 18-24, '20", "Sept 12-20, '20", "Dec 1-7, '20", "Feb 9-14, '21", "May 14-25, '21")) +
  theme_minimal() +
  theme(plot.title = element_text(size = 26),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_line(color = "grey70", size = 0.5),
        panel.grid.major.y = element_line(color = "grey90", size = 0.5),
        axis.title.x = element_text(size = 18),
        axis.text.x = element_text(size = 15),
        axis.title.y = element_text(size = 16),
        legend.title = element_text(size = 12),
        legend.text = element_text(size = 10)) +
  guides(shape = "none")

ggsave("rvd_masks_waves.png", plot = masks_rvd_waves, path = "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Masks", width = 12, height = 8, units = "in")

masks_rvd_waves
rm(all_emms, masks_rvd_waves)
```
