==> code/07-in_dem_cd.Rmd <==
---
title: "05-dems_in_rep_cd"
author: "Christopher Soria"
date: "2024-02-18"
output: html_document
---
models
I'm dropping weights since they no longer apply well after sub-setting the data
Step 1. Do not control for either political ideology nor mask mandates
```{r}
dem_in_rep_concern_model <- lm(binary_concern_strong ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_concern_model <- lm(binary_concern_strong ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + extreme_liberal + extreme_conservative + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_concern_model <- lm(binary_concern_strong ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

print(summary(dem_in_rep_concern_model))
print(summary(rep_in_rep_concern_model))
print(summary(ind_in_rep_concern_model))

CD_PERCENT_DEMOCRAT_range <- seq(min(D_df$CD_PERCENT_DEMOCRAT), max(D_df$CD_PERCENT_DEMOCRAT), length.out = 100)

concern_dem_in_rep_values <- .7 + (dem_in_rep_concern_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_concern_in_rep_values <- .7 + (rep_in_rep_concern_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_concern_in_rep_values <- .7 + (ind_in_rep_concern_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

# Your plotting code
plot(CD_PERCENT_DEMOCRAT_range, concern_dem_in_rep_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Concern",
     xlab = "CD Percent Republican", ylab = "Estimated Concern", ylim = c(0.6, 1), lwd = 2)

# Add the second and third lines using lines()
lines(CD_PERCENT_DEMOCRAT_range, rep_concern_in_rep_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_concern_in_rep_values, type = "l", col = "gray", lwd = 2)

# Add a grid for better readability
grid()

# Add a legend to distinguish the lines
legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red", "gray"), 
       lty = 1, lwd = 2)
```

Step 2. Control for political ideology
```{r}
dem_in_rep_concern_model <- lm(binary_concern_strong ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + extreme_liberal + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_concern_model <- lm(binary_concern_strong ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + extreme_conservative + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_concern_model <- lm(binary_concern_strong ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + extreme_liberal + extreme_conservative + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

CD_PERCENT_DEMOCRAT_range <- seq(min(D_df$CD_PERCENT_DEMOCRAT), max(D_df$CD_PERCENT_DEMOCRAT), length.out = 100)

concern_dem_in_rep_values <- .7 + (dem_in_rep_concern_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_concern_in_rep_values <- .7 + (rep_in_rep_concern_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_concern_in_rep_values <- .7 + (ind_in_rep_concern_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, concern_dem_in_rep_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Concern",
     xlab = "CD Percent Republican", ylab = "Estimated Concern", ylim = c(0.6, 1), lwd = 2)

# Add the second line using lines()
lines(CD_PERCENT_DEMOCRAT_range, rep_concern_in_rep_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_concern_in_rep_values, type = "l", col = "gray", lwd = 2)


# Add a grid for better readability
grid()

# Add a legend to distinguish the lines
legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```

Step 3. Control for political ideology and mask mandates
Neither mask mandates nor political ideology seem to matter in terms of shifting people's stated concern about the spread of the virus (but does make the slopes more similar)
```{r}
#concern
dem_in_rep_concern_model <- lm(binary_concern_strong ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + extreme_liberal + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_concern_model <- lm(binary_concern_strong ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + extreme_conservative + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_concern_model <- lm(binary_concern_strong ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + extreme_liberal + extreme_conservative + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

CD_PERCENT_DEMOCRAT_range <- seq(min(D_df$CD_PERCENT_DEMOCRAT), max(D_df$CD_PERCENT_DEMOCRAT), length.out = 100)

concern_dem_in_rep_values <- .7 + (dem_in_rep_concern_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_concern_in_rep_values <- .7 + (rep_in_rep_concern_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_concern_in_rep_values <- .7 + (ind_in_rep_concern_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range


plot(CD_PERCENT_DEMOCRAT_range, concern_dem_in_rep_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Concern",
     xlab = "CD Percent Republican", ylab = "Estimated Concern", ylim = c(0.6, 1), lwd = 2)

# Add the second line using lines()
lines(CD_PERCENT_DEMOCRAT_range, rep_concern_in_rep_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_concern_in_rep_values, type = "l", col = "gray", lwd = 2)


# Add a grid for better readability
grid()

# Add a legend to distinguish the lines
legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
saving the png
```{r}
output_file_path <- "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Concern/d_cd_concern.png"

png(filename = output_file_path, width = 800, height = 600)

plot(CD_PERCENT_DEMOCRAT_range, concern_dem_in_rep_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Concern",
     xlab = "CD Percent Republican", ylab = "Estimated Concern", ylim = c(0.6, 1), lwd = 2)

# Add the second line using lines()
lines(CD_PERCENT_DEMOCRAT_range, rep_concern_in_rep_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_concern_in_rep_values, type = "l", col = "gray", lwd = 2)


# Add a grid for better readability
grid()

# Add a legend to distinguish the lines
legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)


dev.off()
```

stated contact reduction
Step 1. Do not control for either political ideology not mask mandates 
Steepest slope is independent. Democrats are least likely to change their statement on whether they have reduced contact as republicans in district increase.
```{r}
#contact reduction
dem_in_rep_contact_reduction_model <- lm(contact_reduction ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_contact_reduction_model <- lm(contact_reduction ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_contact_reduction_model <- lm(contact_reduction ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

CD_PERCENT_DEMOCRAT_range <- seq(min(D_df$CD_PERCENT_DEMOCRAT), max(D_df$CD_PERCENT_DEMOCRAT), length.out = 100)

contact_reduction_dem_in_rep_values <- .7 + (dem_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
contact_reduction_rep_in_rep_values <- .7 + (rep_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
contact_reduction_ind_in_rep_values <- .7 + (ind_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, contact_reduction_dem_in_rep_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Contact Reduction",
     xlab = "CD Percent Republican", ylab = "Estimated Contact Reduction", ylim = c(0.6, 1), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, contact_reduction_rep_in_rep_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, contact_reduction_ind_in_rep_values, type = "l", col = "gray", lwd = 2)


grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
Step 2. Control for political ideology
Makes the slope less steep but doesn't change the pattern. 
```{r}
#contact reduction
dem_in_rep_contact_reduction_model <- lm(contact_reduction ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_contact_reduction_model <- lm(contact_reduction ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_contact_reduction_model <- lm(contact_reduction ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

CD_PERCENT_DEMOCRAT_range <- seq(min(D_df$CD_PERCENT_DEMOCRAT), max(D_df$CD_PERCENT_DEMOCRAT), length.out = 100)

contact_reduction_dem_in_rep_values <- .7 + (dem_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
contact_reduction_rep_in_rep_values <- .7 + (rep_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
contact_reduction_ind_in_rep_values <- .7 + (ind_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, contact_reduction_dem_in_rep_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Contact Reduction",
     xlab = "CD Percent Republican", ylab = "Estimated Contact Reduction", ylim = c(0.6, 1), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, contact_reduction_rep_in_rep_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, contact_reduction_ind_in_rep_values, type = "l", col = "gray", lwd = 2)


grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
Step 3. Control for county mask mandates
Steepens the curve and independents are most likely to switch away from saying they greatly reduced their contact as a result of the pandemic as more republicans populate a congressional district. 
```{r}
#contact reduction
dem_in_rep_contact_reduction_model <- lm(contact_reduction ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + county_mask_mandate + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_contact_reduction_model <- lm(contact_reduction ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + county_mask_mandate + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_contact_reduction_model <- lm(contact_reduction ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + county_mask_mandate + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

CD_PERCENT_DEMOCRAT_range <- seq(min(D_df$CD_PERCENT_DEMOCRAT), max(D_df$CD_PERCENT_DEMOCRAT), length.out = 100)

contact_reduction_dem_in_rep_values <- .7 + (dem_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
contact_reduction_rep_in_rep_values <- .7 + (rep_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
contact_reduction_ind_in_rep_values <- .7 + (ind_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, contact_reduction_dem_in_rep_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Contact Reduction",
     xlab = "CD Percent Republican", ylab = "Estimated Contact Reduction", ylim = c(0.6, 1), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, contact_reduction_rep_in_rep_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, contact_reduction_ind_in_rep_values, type = "l", col = "gray", lwd = 2)


grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
saving to png
```{r}
output_file_path <- "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Contact_Reduction/d_cd_contact_reduction.png"

png(filename = output_file_path, width = 800, height = 600)

CD_PERCENT_DEMOCRAT_range <- seq(min(D_df$CD_PERCENT_DEMOCRAT), max(D_df$CD_PERCENT_DEMOCRAT), length.out = 100)

contact_reduction_dem_in_rep_values <- .7 + (dem_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
contact_reduction_rep_in_rep_values <- .7 + (rep_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
contact_reduction_ind_in_rep_values <- .7 + (ind_in_rep_contact_reduction_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, contact_reduction_dem_in_rep_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Contact Reduction",
     xlab = "CD Percent Republican", ylab = "Estimated Contact Reduction", ylim = c(0.6, 1), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, contact_reduction_rep_in_rep_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, contact_reduction_ind_in_rep_values, type = "l", col = "gray", lwd = 2)


grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)

dev.off()
```

Step 1. Do not control for either political ideology not mask mandates 
When I do not control for mask mandates or political ideology, republicans and democrats look pretty similar in how they're impacted by being surrounded by more republicans. Independents have the steepest slop upwards. 
```{r}
##contacts
dem_in_rep_contact_model <- glm.nb(num_cc_nonhh ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` + cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_contact_model <- glm.nb(num_cc_nonhh ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_contact_model <- glm.nb(num_cc_nonhh ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave  + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

CD_PERCENT_DEMOCRAT_range <- seq(min(D_df$CD_PERCENT_DEMOCRAT), max(D_df$CD_PERCENT_DEMOCRAT), length.out = 100)

dem_in_rep_contact_values <- 1 + (dem_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_contact_values <- 1 + (rep_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_contact_values <- 1 + (ind_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_contact_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Non-Household Contacts",
     xlab = "CD Percent Republican", ylab = "Estimated Non-Household Contacts", ylim = c(0.0, 2), lwd = 2)

# Add the second line using lines()
lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_contact_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_contact_values, type = "l", col = "gray", lwd = 2)

print(summary(dem_in_rep_contact_model))
print(summary(rep_in_rep_contact_model))
print(summary(ind_in_rep_contact_model))

grid()

# Add a legend to distinguish the lines
legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
Step 2. Control for political ideology
Controlling for political ideology doesn't change much, but democrats and republicans look more similar
```{r}
dem_in_rep_contact_model <- glm.nb(num_cc_nonhh ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + cityNY +`cityBay Area` + cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_contact_model <- glm.nb(num_cc_nonhh ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_conservative  + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_contact_model <- glm.nb(num_cc_nonhh ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative  + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave  + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

print(summary(dem_in_rep_contact_model))
print(summary(rep_in_rep_contact_model))
print(summary(ind_in_rep_contact_model))

dem_in_rep_contact_values <- 1 + (dem_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_contact_values <- 1 + (rep_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_contact_values <- 1 + (ind_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_contact_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Non-Household Contacts",
     xlab = "CD Percent Republican", ylab = "Estimated Non-Household Contacts", ylim = c(0.0, 2), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_contact_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_contact_values, type = "l", col = "gray", lwd = 2)

grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
Step 3. control for county-level mask mandates
Something funny is going on below. When I control for mask mandates, republican behavior is not at all impacted by how many republicans are around them, but democrats are. This implies that much of the effect of Republicans increasing contacts in higher republican concentrated congressional districts is due to county-level mandates. In other words, regardless of whether a county-mandates exist, democrats and republicans are more likely to interact with non-household contacts when surrounded by more republicans. 
```{r}
dem_in_rep_contact_model <- glm.nb(num_cc_nonhh ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + cityNY +`cityBay Area` + cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_contact_model <- glm.nb(num_cc_nonhh ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_conservative  + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_contact_model <- glm.nb(num_cc_nonhh ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative  + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave  + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

print(summary(dem_in_rep_contact_model))
print(summary(rep_in_rep_contact_model))
print(summary(ind_in_rep_contact_model))

dem_in_rep_contact_values <- 1 + (dem_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_contact_values <- 1 + (rep_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_contact_values <- 1 + (ind_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_contact_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Non-Household Contacts",
     xlab = "CD Percent Republican", ylab = "Estimated Non-Household Contacts", ylim = c(0.0, 2), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_contact_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_contact_values, type = "l", col = "gray", lwd = 2)

grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
saving to png
```{r}
output_file_path <- "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Contacts/d_cd_contacts.png"

png(filename = output_file_path, width = 800, height = 600)

dem_in_rep_contact_values <- 1 + (dem_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_contact_values <- 1 + (rep_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_contact_values <- 1 + (ind_in_rep_contact_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_contact_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Non-Household Contacts",
     xlab = "CD Percent Republican", ylab = "Estimated Non-Household Contacts", ylim = c(0.0, 2), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_contact_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_contact_values, type = "l", col = "gray", lwd = 2)

grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
dev.off()
```
Before I move on, I'm interested in seeing whether there might be a non-linear relationship? No, percent democrats in the district does not move the needle on how republicans react. 

```{r}
rep_in_rep_contact_model <- glm.nb(num_cc_nonhh ~  log_rate_prev_12_01_20 + county_mask_mandate + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race  + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + Categorical_Dem_CD_County_Share, 
                                   data = R_df,
                                   weights = weight_pooled)

print(summary(rep_in_rep_contact_model))
```


Mask usage
Step 1.
```{r}
dem_in_rep_mask_model <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_mask_model <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_mask_model <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

dem_in_rep_mask_values <- .5 + (dem_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_mask_values <- .5 + (rep_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_mask_values <- .5 + (ind_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_mask_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Masked Contacts",
     xlab = "CD Percent Republican", ylab = "Estimated Proportion Masked Contacts", ylim = c(0.4, 1), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_mask_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_mask_values, type = "l", col = "gray", lwd = 2)

grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
Step 2. Control for political ideology
```{r}
dem_in_rep_mask_model <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_mask_model <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_mask_model <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

print(summary(dem_in_rep_mask_model))
print(summary(rep_in_rep_mask_model))
print(summary(ind_in_rep_mask_model))

dem_in_rep_mask_values <- .5 + (dem_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_mask_values <- .5 + (rep_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_mask_values <- .5 + (ind_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_mask_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Masked Contacts",
     xlab = "CD Percent Republican", ylab = "Estimated Proportion Masked Contacts", ylim = c(0.4, 1), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_mask_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_mask_values, type = "l", col = "gray", lwd = 2)

grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
Step 3. control for both poltical ideology and county-level mask mandate
conclusion. Neither political ideology nor mask mandates change the pattern, but both seem to make the R and D slopes appear to be more similar. 
```{r}
dem_in_rep_mask_model <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + county_mask_mandate + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_rep_mask_model <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + county_mask_mandate + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_rep_mask_model <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + county_mask_mandate + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + wave + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

print(summary(dem_in_rep_mask_model))
print(summary(rep_in_rep_mask_model))
print(summary(ind_in_rep_mask_model))

dem_in_rep_mask_values <- .5 + (dem_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_mask_values <- .5 + (rep_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_mask_values <- .5 + (ind_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_mask_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Masked Contacts",
     xlab = "CD Percent Republican", ylab = "Estimated Proportion Masked Contacts", ylim = c(0.4, 1), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_mask_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_mask_values, type = "l", col = "gray", lwd = 2)

grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
saving to png
```{r}
output_file_path <- "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Masks/d_cd_masks.png"

png(filename = output_file_path, width = 800, height = 600)

dem_in_rep_mask_values <- .5 + (dem_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_mask_values <- .5 + (rep_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_mask_values <- .5 + (ind_in_rep_mask_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_mask_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Masked Contacts",
     xlab = "CD Percent Republican", ylab = "Estimated Proportion Masked Contacts", ylim = c(0.4, 1), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_mask_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_mask_values, type = "l", col = "gray", lwd = 2)

grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)

dev.off()
```
Again, let's see if the relationship is non-linear. Looks pretty linear. 

However, when I add the weights, things looks pretty different. 
Let's plot this. 
```{r}
national_R <- national_df %>% 
  filter(political_party == "Republican")

national_D <- national_df %>% 
  filter(political_party == "Democrat")

rep_in_rep_mask_model <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + county_mask_mandate + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + CD_PERCENT_DEMOCRAT, 
                                   data = R_df,
                            weights = weight_pooled)

summary(rep_in_rep_mask_model)

dem_in_rep_mask_model <- lm(Norm_Masks_Used ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + county_mask_mandate + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + CD_PERCENT_REPUBLICAN, 
                            data = D_df,
                            weights = weight_pooled)

summary(dem_in_rep_mask_model)

rm(national_R, national_D)
```



Vaccination
Step 1. no controls
Not significant
```{r}
dem_in_vax_model <- lm(Vaccinated ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_vax_model <- lm(Vaccinated ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_vax_model <- lm(Vaccinated ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

print(summary(dem_in_vax_model))
print(summary(rep_in_vax_model))
print(summary(ind_in_vax_model))

dem_in_rep_vax_values <- .5 + (dem_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_vax_values <- .5 + (rep_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_vax_values <- .5 + (rep_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_mask_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Masked Contacts",
     xlab = "CD Percent Republican", ylab = "Estimated Proportion Masked Contacts", ylim = c(0.4, 0.8), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_mask_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_mask_values, type = "l", col = "gray", lwd = 2)

grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```

Step 2. 
Not significant
```{r}
dem_in_vax_model <- lm(Vaccinated ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_vax_model <- lm(Vaccinated ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_vax_model <- lm(Vaccinated ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

print(summary(dem_in_vax_model))
print(summary(rep_in_vax_model))
print(summary(ind_in_vax_model))

dem_in_rep_vax_values <- .5 + (dem_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_vax_values <- .5 + (rep_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_vax_values <- .5 + (rep_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_mask_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Vaccination",
     xlab = "CD Percent Republican", ylab = "Estimated Probability of Vaccination", ylim = c(0.4, 0.8), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_mask_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_mask_values, type = "l", col = "gray", lwd = 2)

grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
Step 3. all controls
not significant
```{r}
dem_in_vax_model <- lm(Vaccinated ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + county_mask_mandate + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + CD_PERCENT_DEMOCRAT, 
                                   data = D_df)

rep_in_vax_model <- lm(Vaccinated ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + county_mask_mandate + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + CD_PERCENT_DEMOCRAT, 
                                   data = R_df)

ind_in_vax_model <- lm(Vaccinated ~  log_rate_prev_12_01_20 + COUNTY_RUCC_2013 + county_mask_mandate + male + educ_group + resp_yob + r_working + resp_hhsize + r_race + extreme_liberal + extreme_conservative + cityNY +`cityBay Area` +cityPhiladelphia + cityPhoenix +cityBoston + cityAtlanta + CD_PERCENT_DEMOCRAT, 
                                   data = I_df)

print(summary(dem_in_vax_model))
print(summary(rep_in_vax_model))
print(summary(ind_in_vax_model))

dem_in_rep_vax_values <- .5 + (dem_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_vax_values <- .5 + (rep_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_vax_values <- .5 + (rep_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_mask_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Masked Contacts",
     xlab = "CD Percent Republican", ylab = "Estimated Proportion Masked Contacts", ylim = c(0.4, 0.8), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_mask_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_mask_values, type = "l", col = "gray", lwd = 2)

grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)
```
save to png
```{r}
output_file_path <- "/Users/chrissoria/Documents/Research/BICS_Political_Polarization/plots/Vaccines/d_cd_vaccines.png"

png(filename = output_file_path, width = 800, height = 600)

dem_in_rep_vax_values <- .5 + (dem_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
rep_in_rep_vax_values <- .5 + (rep_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range
ind_in_rep_vax_values <- .5 + (rep_in_vax_model$coefficients["CD_PERCENT_DEMOCRAT"]) * CD_PERCENT_DEMOCRAT_range

plot(CD_PERCENT_DEMOCRAT_range, dem_in_rep_mask_values, type = "l", col = "blue",
     main = "Estimated Relationship between CD Percent Republican and Masked Contacts",
     xlab = "CD Percent Republican", ylab = "Estimated Proportion Masked Contacts", ylim = c(0.4, 0.8), lwd = 2)

lines(CD_PERCENT_DEMOCRAT_range, rep_in_rep_mask_values, type = "l", col = "red", lwd = 2)
lines(CD_PERCENT_DEMOCRAT_range, ind_in_rep_mask_values, type = "l", col = "gray", lwd = 2)

grid()

legend("topright", 
       legend = c("Democrats", "Republicans", "Independents"), 
       col = c("blue", "red","gray"), 
       lty = 1, lwd = 2)

dev.off()
```

```{r}
rm(dem_in_vax_model,dem_in_vax_model,ind_in_vax_model,dem_in_rep_vax_values,rep_in_rep_vax_values,ind_in_rep_vax_values)
rm(dem_in_rep_mask_model,rep_in_rep_mask_model,ind_in_rep_mask_model,dem_in_rep_mask_values,rep_in_rep_mask_values,ind_in_rep_mask_values)
rm(dem_in_rep_contact_model,rep_in_rep_contact_model,ind_in_rep_contact_model,dem_in_rep_contact_values,rep_in_rep_contact_values,ind_in_rep_contact_values)
rm(dem_in_rep_concern_model,rep_in_rep_concern_model,ind_in_rep_concern_model,concern_dem_in_rep_values,rep_concern_in_rep_values,ind_concern_in_rep_values)
rm(dem_in_rep_contact_model,rep_in_rep_contact_model,rep_in_vax_model,rvd_masks_model,ind_in_vax_model,ind_in_rep_contact_model,dem_in_vax_model,ind_in_rep_mask_model)
rm(dem_in_rep_contact_model,ind_in_rep_mask_model, rep_in_rep_contact_model, rep_in_rep_contact_reduction_model, ind_in_rep_contact_reduction_model, dem_in_rep_contact_reduction_model)
rm(CD_PERCENT_DEMOCRAT_range,cd_percent_republican_range,concern_in_rep_values,contact_reduction_rep_in_rep_values)
rm(contact_reduction_ind_in_rep_values, contact_reduction_dem_in_rep_values,contact_reduction_dem_in_rep_values, contact_reduction_ind_in_rep_values,output_file_path)
```
